% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplist.R
\name{add.cplist}
\alias{add.cplist}
\title{Function to add entries to cplist}
\usage{
\method{add}{cplist}(cplist, new.j, new.k, newentry)
}
\arguments{
\item{cplist}{list containing an n x 3 matrix, and the index of the last
nonempty (i.e. not all NA's) row.}
}
\description{
Function to add entries to cplist
}
