% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{add.intervals}
\alias{add.intervals}
\title{Add a single interval to existing set of intervals}
\usage{
\method{add}{intervals}(intervals, new.s, new.e)
}
\arguments{
\item{new.s}{new start point}

\item{new.e}{new end point}

\item{old.intervals}{an object of class |intervals|}
}
\description{
Add a single interval to existing set of intervals
}
