% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{add2.intervals}
\alias{add2.intervals}
\title{Add winning intervals from the original object, to the |intervals| object.}
\usage{
\method{add2}{intervals}(intervals, winning.wbs.obj, stop.time = NULL)
}
\description{
Add winning intervals from the original object, to the |intervals| object.
}
