% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbs-helper.R
\name{addd}
\alias{addd}
\title{Adds a matrix containing information about wbs selection even, as a last
element to an existing list of such matrices. the name of this new element
is of the form "j,k". this is in lieu of |cplist| which was used to store
information for sbs-fixed-threshold events. also note: more than one
maximizer per branch is /fine/!}
\usage{
addd(mylist, j, k, semat)
}
\arguments{
\item{mylist}{a list of matrices}
}
\description{
Adds a matrix containing information about wbs selection even, as a last
element to an existing list of such matrices. the name of this new element
is of the form "j,k". this is in lieu of |cplist| which was used to store
information for sbs-fixed-threshold events. also note: more than one
maximizer per branch is /fine/!
}
