% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbs-helper.R
\name{adddd}
\alias{adddd}
\title{Make an addition to the sign list. Specifically, adds elements}
\usage{
adddd(newsigns, M, env)
}
\arguments{
\item{newsigns}{the signs to add.}

\item{M}{the indices in env$intervals}

\item{env}{environment where signs are embedded in}
}
\description{
Make an addition to the sign list. Specifically, adds elements
}
