% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addpv.R
\name{addpv.bsfs}
\alias{addpv.bsfs}
\title{Appends the inference results to an object of class |bsFs|.}
\usage{
\method{addpv}{bsfs}(obj, loc = NULL, type = c("plain", "addnoise"),
  sigma, sigma.add = NULL, declutter = FALSE, min.num.things = 30,
  max.numIS = 2000, mn = NULL, only.test.nulls = FALSE,
  bootsub = FALSE, nboot = 10000, v2 = FALSE)
}
\arguments{
\item{obj}{object of type bsFs.}

\item{loc}{only test locations in \code{loc}.}

\item{type}{One of \code{ c("plain", "addnoise")}. If equal to
\code{"addnoise"}, then \code{sigma.add} needs to be
provided. Currently, the polyhedron is not formed (due to memory
concerns) but instead it uses the \code{type=pre-multiply} option.}

\item{sigma}{Noise level (standard deviation) of data.}

\item{sigma.add}{Additive noise. Defaults to NULL, in which case no additive
noise randomization inference is done.}

\item{mn}{Original mean vector. This is purely for simulation purposes,
along with the |only.test.nulls| option.}

\item{only.test.nulls}{If \code{TRUE}, only test the contrasts whose true
mean is zero (i.e. the ones that constitute null tests).}

\item{inference.type}{One of \code{c("rows","pre-multiply")}. Defaults to '
\code{"rows"}. Use \code{"pre-multiply"} if the polyhedron is too big
for memory.  (Warning: the rest is more of a note to self (Justin) than
' for users.). The use of \code{"pre-multiply"} was originally built for
' WBS. It actually prevents \code{polyhedra.wbsfs()} from having to form
' the entire WBS polyhedron in the first place. It also makes importance
' sampling faster the \code{type="rows"} option for WBS (thanks to
manual ' speedups we've made), but slower for all other segmentation
methods ' since there are no such manual tweaks for speedup. The most
crucial ' difference for the user is perhaps that, when the size of the
polyhedron ' is too big in memory, then this is a !necessity! as it
circumvents ' having to actually form the polyhedron.}
}
\description{
Appends the inference results to an object of class |bsFs|.
}
