% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addpv.R
\name{addpv.cbsfs}
\alias{addpv.cbsfs}
\title{Appends the inference results to an object of class |bsFs|.}
\usage{
\method{addpv}{cbsfs}(obj, loc = NULL, type = c("plain", "addnoise"),
  sigma, sigma.add = NULL, declutter = FALSE, mn = NULL,
  min.num.things = 30, numIntervals = NULL,
  inference.type = c("rows", "pre-multiply"))
}
\arguments{
\item{obj}{object of type bsFs}

\item{loc}{only test locations in \code{loc}.}

\item{type}{One of \code{ c("plain", "addnoise")}. If equal to
\code{"addnoise"}, then \code{sigma.add} needs to be provided.}

\item{sigma}{Noise level (standard deviation) of data.}

\item{sigma.add}{Additive noise. Defaults to NULL, in which case no additive
noise randomization inference is done.}

\item{mn}{original mean vector.}

\item{inference.type}{One of \code{c("rows","pre-multiply")}. Defaults to '
\code{"rows"}. Use \code{"pre-multiply"} if the polyhedron is too big
for memory.}
}
\description{
Appends the inference results to an object of class |bsFs|.
}
