% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addpv.R
\name{addpv.wbsfs}
\alias{addpv.wbsfs}
\title{Appends the inference results to an object of class |bsFs|.}
\usage{
\method{addpv}{wbsfs}(obj, locs = NULL, type = c("plain", "rand"),
  sigma, declutter = FALSE, mn = NULL, min.num.things = 30,
  sigma.add = NULL, max.numIS = 5000, numIS.base = 10,
  only.test.nulls = FALSE, verbose = FALSE, vlist = NULL,
  mc.cores = 1, inference.type = c("pre-multiply", "rows"))
}
\arguments{
\item{obj}{object of type bsFs}

\item{locs}{only test locations in \code{locs}.}

\item{type}{One of \code{ c("plain", "addnoise")}. If equal to
\code{"addnoise"}, then \code{sigma.add} needs to be provided.}

\item{sigma}{Noise level (standard deviation) of data.}

\item{mn}{original mean vector.}

\item{max.numIS}{Maximum number of importance sampling replicates to perform.}

\item{inference.type}{One of \code{c("pre-multiply")}. Defaults to '
\code{"pre-multiply"}. Use \code{"pre-multiply"} if the polyhedron is
too big for memory. There is really no reason to use \code{"rows"} in
WBS.}
}
\description{
Appends the inference results to an object of class |bsFs|.
}
