% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplist.R
\name{addrow.cplist2}
\alias{addrow.cplist2}
\title{Function to add entries to cplist}
\usage{
\method{addrow}{cplist2}(cplist2, j = NA, k = NA, s = NA, e = NA,
  s0 = NA, e0 = NA, z = NA, pass = NA)
}
\arguments{
\item{cplist2}{list containing an n x 3 matrix, and the index of the last
nonempty (i.e. not all NA's) row.}
}
\description{
Function to add entries to cplist
}
