% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{binary_search}
\alias{binary_search}
\title{Binary search an integer vector for goal Returns index of goal, in the
vector.}
\usage{
binary_search(vec, goal, verbose = FALSE)
}
\description{
Binary search an integer vector for goal Returns index of goal, in the
vector.
}

