% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binseg.R
\name{binseg.by.size}
\alias{binseg.by.size}
\title{Function to carry out fixed-number-of-steps binary segmentation.}
\usage{
binseg.by.size(y, numsteps, verbose = FALSE)
}
\arguments{
\item{y}{numeric vector, data}

\item{numsteps}{desired number of changepoints}

\item{verbose}{set to true for algorithm run details.}
}
\description{
Function to carry out fixed-number-of-steps binary segmentation.
}

