% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsft.R
\name{binseg.by.thresh.inner}
\alias{binseg.by.thresh.inner}
\title{Inner function for binary segmentation with fixed threshold. The wrapper
\code{binseg.by.thresh()} is intended to be used by user. Note, when
\code{thresh} is set to zero, then this can be used to collect the
unbalanced haar wavelet basis.}
\usage{
binseg.by.thresh.inner(y, thresh, s = 1, e = length(y), j = 1,
  k = 1, verbose = F, env = NULL)
}
\arguments{
\item{y}{The original data.}

\item{thresh}{Threshold for \deqn{\tilde{X}}. This serves as a stopping rule
for the recursion.}

\item{s}{Starting index, in the vector-valued data. Must be an integer
larger than or equal to 1, and strictly smaller than \code{e}.}

\item{e}{Ending index, in the vector-valued data. Must be an integer smaller
than or equal to \code{n}, and strictly larger than \code{s}.}

\item{j}{The depth of the recursion on hand.}

\item{k}{The indexing of the node location, from left to right, in the
\emph{complete} binary tree.}

\item{n}{The length of the data \code{y}.}
}
\description{
Inner function for binary segmentation with fixed threshold. The wrapper
\code{binseg.by.thresh()} is intended to be used by user. Note, when
\code{thresh} is set to zero, then this can be used to collect the
unbalanced haar wavelet basis.
}
