% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binSeg_fixedSteps.R
\name{binSeg_fixedSteps}
\alias{binSeg_fixedSteps}
\title{Binary segmentation with fixed steps}
\usage{
binSeg_fixedSteps(y, numSteps)
}
\arguments{
\item{y}{numeric vector to contain data}

\item{numSteps}{numeric of number of steps}
}
\value{
a bsFs object
}
\description{
y must not have duplicated values. This is to avoid
degenerate behavior of binary segmentation
}

