% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binseginf-package.R
\docType{package}
\name{binseginf-package}
\alias{binseginf-package}
\title{Dummy package to test out C++}
\description{
Some boring description here
}
\author{
Sangwon Hyun \email{shyun@cmu.edu}
}
\keyword{package}
