% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{bootstrap_sample}
\alias{bootstrap_sample}
\title{Creates bootstrap sample of numeric vector \code{vec}, optionally with a
seed.}
\usage{
bootstrap_sample(vec, size = length(vec), seed = NULL)
}
\arguments{
\item{vec}{Numeric vector}

\item{seed}{Seed for random number generation}
}
\value{
resampled \code{vec}.
}
\description{
Creates bootstrap sample of numeric vector \code{vec}, optionally with a
seed.
}
