% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue.R
\name{bryc.tnorm.surv}
\alias{bryc.tnorm.surv}
\title{A numerical approximation, to be used as a substitute for
\code{mpfr.tnorm.surv()}.}
\usage{
bryc.tnorm.surv(z, mean = 0, sd = 1, a, b)
}
\description{
A numerical approximation, to be used as a substitute for
\code{mpfr.tnorm.surv()}.
}
