% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsfs.R
\name{bsfs}
\alias{bsfs}
\title{Binary segmentation with fixed steps
y must not have duplicated values. This is to avoid
degenerate behavior of binary segmentation}
\usage{
bsfs(y, numSteps, sigma.add = NULL, numIntervals = NULL,
  ic.stop = FALSE, y.addnoise = NULL)
}
\arguments{
\item{y}{numeric vector to contain data}

\item{numSteps}{numeric of number of steps}

\item{sigma.add}{is the amount (standard deviation) of i.i.d. Gaussian noise
added to the data.}

\item{y.addnoise}{Manully inputted additive noise. Defaults to \code{NULL}.}
}
\value{
a bsfs object, which is a list of information regarding the fitted
    algorithm. The list component \code{y} is the data used for actual
    fitting; \code{y.orig} is the pre-noise original data; \code{y.addnoise}
    (if not null) is the added noise.
}
\description{
Binary segmentation with fixed steps
y must not have duplicated values. This is to avoid
degenerate behavior of binary segmentation
}
