% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbsfs.R
\name{cbsfs}
\alias{cbsfs}
\title{Circular binary segmentation - fixed steps}
\usage{
cbsfs(y, numSteps, sigma.add = NULL, numIntervals = NULL,
  y.addnoise = NULL)
}
\arguments{
\item{y}{data, numeric vector}

\item{numSteps}{integer for number of jumps}

\item{sigma.add}{is the amount (standard deviation) of i.i.d. Gaussian noise
added to the data.}
}
\value{
cbs object, which is a list of information regarding the fitted
    algorithm. The list component \code{y} is the data used for actual
    fitting; \code{y.orig} is the pre-noise original data; \code{y.addnoise}
    (if not null) is the added noise.
}
\description{
Returns a list containing a tree denoting how the input \code{y} is
segmented and the input \code{numSteps}. If the output is stored
in a variable \code{res}, use \code{jumps(res)} to enumerate all the
jump locations.
}
