% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{check.orth.basis}
\alias{check.orth.basis}
\title{Function to check if the columns in \code{basislist} are orthonormal.}
\usage{
check.orth.basis(basislist, tol = 1e-10)
}
\arguments{
\item{tol}{Numerical allowance; letting inner products to be up to size
\code{tol}.}
}
\description{
Function to check if the columns in \code{basislist} are orthonormal.
}

