% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbsft-poly.R
\name{check_polyhedra}
\alias{check_polyhedra}
\title{Temporary function to check if it is correct; NOT to be called at
runtime, but only at test time or internally.}
\usage{
check_polyhedra(poly, y)
}
\arguments{
\item{poly}{An object of class polyhedra}

\item{y}{A data vector with the appropriate length, used when
creating this polyhedron.}
}
\description{
Temporary function to check if it is correct; NOT to be called at
runtime, but only at test time or internally.
}
