% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circularBinSeg_fixedSteps.R
\name{circularBinSeg_fixedSteps}
\alias{circularBinSeg_fixedSteps}
\title{Circular binary segmentation - fixed steps}
\usage{
circularBinSeg_fixedSteps(y, numSteps)
}
\arguments{
\item{y}{data, numeric vector}

\item{numSteps}{integer for number of jumps}
}
\value{
cbs object
}
\description{
Returns a list containing a tree denoting how the input \code{y} is
segmented and the input \code{numSteps}. If the output is stored 
in a variable \code{res}, use \code{jumps(res)} to enumerate all the
jump locations.
}
