% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplist.R
\name{collapse}
\alias{collapse}
\title{Helper to collapse matrix (with NAs) to a vector of unique elements.}
\usage{
collapse(mat)
}
\description{
Helper to collapse matrix (with NAs) to a vector of unique elements.
}
