% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyhedra.R
\name{combine.polyhedra}
\alias{combine.polyhedra}
\title{Combines several polyhedra to a single polyhedron. It is possible to add
\code{NULL} valued objects with other polyhedra (to no effect)}
\usage{
\method{combine}{polyhedra}(...)
}
\arguments{
\item{...}{Polyhedra objects to add. It is permissible for some of these to
be \code{NULL} valued.}
}
\description{
Combines several polyhedra to a single polyhedron. It is possible to add
\code{NULL} valued objects with other polyhedra (to no effect)
}
