% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyhedra.R
\name{combine}
\alias{combine}
\title{Generic for functions that combine same-type things.}
\usage{
combine(obj, ...)
}
\arguments{
\item{obj}{\code{polyhedra} class object.}
}
\description{
Generic for functions that combine same-type things.
}
