% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_splitting.R
\name{confidence_interval_ss}
\alias{confidence_interval_ss}
\title{Confidence interval for sample splitting}
\usage{
confidence_interval_ss(y, contrast, alpha = 0.95)
}
\arguments{
\item{y}{numeric vector}

\item{contrast}{contrast numeric vector}

\item{alpha}{numeric between 0 and 1 with default of 0.95. This is the
significance level, guaranteeing that alpha percentage
of the intervals will cover the true parameter.}
}
\description{
This is limited to two-sided confidence intervals currently
}
