% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyhedra.R
\name{contained.polyhedra}
\alias{contained.polyhedra}
\title{Check if y is in polyhedra.}
\usage{
\method{contained}{polyhedra}(obj, y)
}
\arguments{
\item{obj}{\code{polyhedra} class object.}
}
\description{
Check if y is in polyhedra.
}
