% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyhedra.R
\name{contained}
\alias{contained}
\title{Check if y is in polyhedra (generic).}
\usage{
contained(obj, ...)
}
\arguments{
\item{obj}{\code{polyhedra} class object.}
}
\description{
Check if y is in polyhedra (generic).
}
