% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_vector.R
\name{contrast_vector.bsFs}
\alias{contrast_vector.bsFs}
\title{Generate contrast bector}
\usage{
\method{contrast_vector}{bsFs}(obj, jump.idx, sorted = F,
  type = c("segment", "spike"), ...)
}
\arguments{
\item{obj}{bsFs object}

\item{jump.idx}{index among the list of jumps to jump at}

\item{sorted}{boolean on whether or not jumps should be sorted}

\item{type}{segment or spike}

\item{...}{not used}
}
\value{
numeric vector
}
\description{
Generate contrast bector
}

