% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{contrast_vector}
\alias{contrast_vector}
\title{Generic function to form contrast vectors}
\usage{
contrast_vector(obj, ...)
}
\arguments{
\item{obj}{object}

\item{...}{additional parameters}
}
\value{
vector of numerics
}
\description{
Generic function to form contrast vectors
}

