% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_splitting.R
\name{contrast_vector_ss}
\alias{contrast_vector_ss}
\title{Generate contrast bector}
\usage{
contrast_vector_ss(obj, jump.idx, sorted = F)
}
\arguments{
\item{obj}{object}

\item{jump.idx}{index among the list of jumps to jump at}

\item{sorted}{boolean on whether or not jumps should be sorted}
}
\value{
numeric vector
}
\description{
Only does segment tests
}
