% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplist.R
\name{cplist}
\alias{cplist}
\title{Constructor for cplist object.}
\usage{
cplist(nrow)
}
\arguments{
\item{nrow}{creates an all-NA matrix of dimension nrow x 3. The first two
columns must be the numeric (no check yet), but the last column can be
of any type you want. Initializes to numeric.}
}
\description{
Constructor for cplist object.
}
