% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplist.R
\name{cplist2}
\alias{cplist2}
\title{Constructor for cplist2 object. Used for CBS-fixed-threshold.}
\usage{
cplist2(nrow)
}
\arguments{
\item{nrow}{creates an all-NA matrix of dimension nrow x 8. Each row
contains \code{(j,k,s,e,s0,e0,z,pass)}.}
}
\description{
Constructor for cplist2 object. Used for CBS-fixed-threshold.
}
