% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{CpVector}
\alias{CpVector}
\title{CpVector Class Constructor}
\usage{
CpVector(n, jump.height = 0, jump.loc = NA, func = stats::rnorm, ...)
}
\arguments{
\item{n}{the number of realizations}

\item{jump.height}{a numeric vector of the jump heights}

\item{jump.loc}{a numeric vector (between 0 (inclusive) and 1 (exclusive))
of the jump locations between 0 and 1.}

\item{func}{the function to dictate the noise model. The first parameter
of func must dictate how many samples are drawn.}

\item{...}{additional parameters to pass into func.}
}
\value{
a CpVector instance
}
\description{
Creates an instance of the CpVector, short for "Change Point Vector".
A CpVector objects has 3 elements, data (the realization from the
changepoint model), jump.height (the height of jumps in the true signal)
and jump.loc (the locations of the jumps in the true signal).
}
\details{
Here, CpVector is draws samples from a piecewise constant true signal from
points 1 to n. The jump.loc is a vector between 0 and 1 (inclusive and 
exclusive repsectively), and CpVector automatically scales jump.loc to
fit between 1 and n. Here, jump.loc must always be one less element than
jump.height.

Also, n must be large enough with respect to jump.loc that round(n*jump.loc)
must be distinct integers. Otherwise, an error will be thrown.

Here, func is a function that dictates the noise model. For example, the
default is rnorm.
}
