% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{cusum}
\alias{cusum}
\title{Computes the CUSUM (cumulative sum) statistic.}
\usage{
cusum(s, b, e, n = NULL, y = NULL, right.to.left = TRUE,
  contrast.vec = FALSE, unsigned = FALSE, sign.of.vy = NULL)
}
\arguments{
\item{s}{starting index.}

\item{b}{breakpoint index.}

\item{e}{end index.}

\item{n}{length of data. Defaults to \code{length(y)}.}

\item{y}{data.}

\item{right.to.left}{Whether you want right-to-left difference in the cusum
calculation. Defaults to TRUE.}

\item{contrast.vec}{If TRUE, then the contrast vector v for cusum=v'y is
returned.}

\item{unsigned}{if TRUE, then returns the contrast vector that makes the
contrast with |y| positive, or the absolute value of the contrast.}
}
\description{
Note, we calculate this as the right-to-left difference, by default.
}
