% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{cv.bsfs}
\alias{cv.bsfs}
\title{Train a binseg changepoint model size using cross validation.}
\usage{
cv.bsfs(y, max.numSteps = 30, numsplit = 2)
}
\description{
Train a binseg changepoint model size using cross validation.
}
