% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{cv2.bsfs}
\alias{cv2.bsfs}
\title{(A copy for debugging) Train a binseg changepoint model size using cross
validation.}
\usage{
cv2.bsfs(y, max.numSteps = 30, numsplit = 2)
}
\description{
(A copy for debugging) Train a binseg changepoint model size using cross
validation.
}
