% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{declutter}
\alias{declutter}
\title{Helper function to take all neighboring-to-each-other clusters, and
declutter them by removing all but middle value of centroids. The sign is
only returned if the signs all agree.}
\usage{
declutter(coords, coords.sign, how.close = 1)
}
\arguments{
\item{coords}{(not necessarily sorted) Coordinates.}

\item{coords.sign}{Accompanying signs.}

\item{how.close}{how close you want the cluster members to be.}
}
\value{
List containing |cp| and |cp.sign|, which are lists containing the
    de-cluttered changepoints and signs. Additionally, the list contains the
    raw, uncluttered versions |cp.raw| and |cp.sign.raw|.
}
\description{
Helper function to take all neighboring-to-each-other clusters, and
declutter them by removing all but middle value of centroids. The sign is
only returned if the signs all agree.
}
