% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{declutter_new}
\alias{declutter_new}
\title{Helper function to take all neighboring-to-each-other clusters,
And declutter them by removing all but (rounded up) centroids}
\usage{
declutter_new(coords, coords.sign = NULL, how.close = 1)
}
\arguments{
\item{coords.sign}{optional set of signs.}
}
\description{
Helper function to take all neighboring-to-each-other clusters,
And declutter them by removing all but (rounded up) centroids
}
