% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{defaultColors}
\alias{defaultColors}
\title{Default Colors}
\usage{
defaultColors(vec = NA, transparency = 1)
}
\arguments{
\item{vec}{a vector of integers ranging between 
.defaultColorRange() (inclusive)}

\item{transparency}{a numeric between 0 and 1 (inclusive) to set the
transparency of the colors}
}
\value{
void
}
\description{
Default Colors
}
