% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{defaultPlotDefaults}
\alias{defaultPlotDefaults}
\title{Default Plot Settings}
\usage{
defaultPlotDefaults()
}
\value{
void
}
\description{
Removes all the plots and currently rendering pdf figures and sets
all the default graphical parameters. Useful for standardizing the figures.
}
\details{
This sets the graphical parameters, "pch", "cex", "family" and "mar".
}
