% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbsfs.R
\name{.cusum_cbs}
\alias{.cusum_cbs}
\title{Compute the circular cusum statistcs}
\usage{
.cusum_cbs(x, vec)
}
\arguments{
\item{x}{vector of 2 integers.}

\item{vec}{numeric vector}
}
\value{
a value
}
\description{
\code{x} are the breakpoints (vector of length 2), and \code{vec}
denotes the vector of cumulative sums. Output the circular cusum
statistic for this proposed split.
}
\details{
\code{x} must be 2 integers where 1) the first element must be
larger than or equal to 1, 2) the second element must be smaller than
or equal to \code{length(vec)}, 3) the first element must be
smaller than or equal to the second element, and 4) the first element
cannot be 1 when the last element is \code{length(vec)}.
}
