% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbsfs-poly.R
\name{.cusum_cbs_contrast_full}
\alias{.cusum_cbs_contrast_full}
\title{Form contrast vector for circular binary segmentation}
\usage{
.cusum_cbs_contrast_full(start, idx, end, n)
}
\arguments{
\item{start}{integer denoting the first (left) index of the left shoulder}

\item{idx}{a vector of 2 integers containing the first (left) index
of the hump and the last (right) index of the hump}

\item{end}{integer denoting the last (right) index of the right shoulder}

\item{n}{length of the contrast vector}
}
\value{
a vector
}
\description{
The vector is made such that for a data vector \code{y}, if the
output of \code{.cusum_cbs_contrast_full} is \code{res}, then
\code{res \%*\% y} is the circular cusum statistic for a particular
set of indices.
}
\details{
Here, \code{start}, \code{idx}, \code{end} are all integers between 1
and \code{n}
}
