% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbsfs.R
\name{.enumerate_breakpoints_cbs}
\alias{.enumerate_breakpoints_cbs}
\title{Enumerate all the possible jump locations for circular binary segmentation}
\usage{
.enumerate_breakpoints_cbs(n, start = 1)
}
\arguments{
\item{n}{integer}

\item{start}{integer}
}
\value{
a matrix
}
\description{
Enumerates all the possible places the start and end of the hump
can occur for a segment of length \code{n}. This is essentially the same
as \code{t(cbind(combn(1:(n-1), 2), rbind(2:(n-1), rep(n, n-2)), rbind(1:n, 1:n)))}.
The first part, \code{combn(1:(n-1), 2)}, accounts for any humps with
a right shoulder and length larger than 1, possibly no left shoulder.
The second part, \code{rbind(2:n, rep(n, n-1)))} accounts for any humps with no
right shoulder and length larger than 1, but has a left shoulder.
The third part, \code{rbind(1:n, 1:n)}
accounts for humps of length 1.
}
\details{
\code{start} shifts all the indices appropriately to the right.
}
