% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbsfs.R
\name{.find_breakpoint_cbs}
\alias{.find_breakpoint_cbs}
\title{Find breakpoints for circular binary segmentation}
\usage{
.find_breakpoint_cbs(y)
}
\arguments{
\item{y}{numeric vector}
}
\value{
a list
}
\description{
Given a vector \code{y} (representating a segment of the data vector),
find the best place to split according to circular binary segmentation.
It returns a list with \code{breakpoint}, a vector of length 2 (for the
start and end of the hump), and \code{cusum} for the circular
cusum statistic.
}
\details{
Here, the \code{breakpoint} is encoded so \code{breakpoint[1]-1}
is the last index of left shoulder (if larger than 0), and
\code{breakpoint[2]+1} is the first index of the right shoulder.
}
