% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_matrix_cbs.R
\name{.form_comparison_cbs}
\alias{.form_comparison_cbs}
\title{Form comparisons for circular binary segmentation}
\usage{
.form_comparison_cbs(tree, nodeName, breakpoint, excluding = T)
}
\arguments{
\item{tree}{tree object}

\item{nodeName}{node name to split on right now}

\item{breakpoint}{vector of 2 positive integers where the split will occur}

\item{excluding}{boolean}
}
\value{
a list
}
\description{
For a particular node in \code{nodeName} (for example, \code{"1-10"}) and
a desired \code{breakpoint} (for example \code{c(3,5)}), form matrices
with 4 columns on all the possible comparisons made.
}
\details{
The returned object has a \code{winning} element, a matrix with 1 row and
4 columns. It denotes where the split dictated by \code{breakpoint} occurs,
in this example, \code{matrix(c(1,3,5,10), nrow = 1)}. 

The returned object also has a \code{losing} element, a matrix with 4 columns.
Each row denotes all the other splits that could have occurred that is not
in \code{winning}.

\code{excluding} is a boolean. If \code{TRUE}, then the comparisons will
include other nodes in \code{losing}. Otherwise, it will not. Also, \code{breakpoint}
can be \code{NA}.
}
