% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_matrix_cbs.R
\name{.fourColumnMatrix_from_nodeMatrix}
\alias{.fourColumnMatrix_from_nodeMatrix}
\title{Form comparisons from a matrix}
\usage{
.fourColumnMatrix_from_nodeMatrix(mat)
}
\arguments{
\item{mat}{a matrix with 2 columns, where each row denotes the left (first)
index of the left shoulder of some segment, and the right (last) index of the
right shoulder of the same segment.}
}
\value{
a matrix
}
\description{
Form the comparisons in a 4-column matrix
}
