% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_matrix_cbs.R
\name{.fourColumnMatrix_from_nodeVec}
\alias{.fourColumnMatrix_from_nodeVec}
\title{Form comparisons from a vector}
\usage{
.fourColumnMatrix_from_nodeVec(vec, exclude = NA)
}
\arguments{
\item{vec}{vector of 2 integers}

\item{exclude}{\code{NA} or a vector of 2 integers}
}
\value{
a matrix
}
\description{
This function returns a 4-column matrix of positive integers.
}
\details{
\code{vec} is two positive integers in increasing order. This
returns all the possible circular binary segmentation comparisons
where \code{vec[1]} is the left (first) index of the left shoulder, and \code{vec[2]} 
is the right (last) index of the right shoulder.

\code{exclude} is an optional argument. If included (i.e., not \code{NA}),
it should be a vector of 2 positive integers and
the returned matrix will omit the row including \code{exclude}.
}
