% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs-poly.R
\name{.gammaRows_from_comparisons}
\alias{.gammaRows_from_comparisons}
\title{Helper to make a set of rows (or multiplications with v), for a winning
location and sign, at a particular step of the binary segmentation
algorithm.}
\usage{
.gammaRows_from_comparisons(
  vec,
  mat,
  sign.win,
  n,
  y = NULL,
  v = NULL,
  w = NULL,
  inference.type = c("rows", "pre-multiply"),
  excessive = FALSE
)
}
\arguments{
\item{vec}{contains the start, break, and endpoints of this winner.}

\item{mat}{is a 3-column matrix that contains the start, break and endpoint
of all the losers at that stage.}

\item{sign.win}{winning sign of the winner's CUSUM.}

\item{excessive}{Experimental feature that /additionally/ (and excessively)
conditions on the signs of competitor CUSUMs.}
}
\description{
Helper to make a set of rows (or multiplications with v), for a winning
location and sign, at a particular step of the binary segmentation
algorithm.
}
