% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbsfs-poly.R
\name{.gammaRows_from_comparisons_cbsfs}
\alias{.gammaRows_from_comparisons_cbsfs}
\title{Form the lines of the gamma matrix}
\usage{
.gammaRows_from_comparisons_cbsfs(vec, mat, sign_win, n, add = F)
}
\arguments{
\item{vec}{a vector of integers}

\item{mat}{a matrix of integers}

\item{sign_win}{1 or -1, denoting the sign of the jump denoted in \code{vec}}

\item{n}{number of elements}

\item{add}{boolean}
}
\value{
a matrix
}
\description{
\code{vec} is a vector of 4 positive integers less than \code{n}, where
(from left to right) denote the first (left) index of the left shoulder,
the first (left) index of the hump, the last (right) index of the hump,
and the last (right) index of the right shoulder. Similarly,
\code{mat} is a matrix with 4 columns.
}
\details{
Typically, this function is used such that the circular cusum
statistic formed with the indices in \code{vec} is larger than the
circular cusum statistic formed with the indicies in any of the rows of
\code{mat}.

This function returns a matrix with \code{n} columns, and \code{2*nrow(mat)}
rows.
}
