% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbsfs.R
\name{.get_max_info.wbsfs}
\alias{.get_max_info.wbsfs}
\title{Helper to /manually/ obtain maximizing information from a |wbs| class object, at i'th step}
\usage{
\method{.get_max_info}{wbsfs}(wbs.obj, istep)
}
\description{
Helper to /manually/ obtain maximizing information from a |wbs| class object, at i'th step
}
