% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbsft-poly.R
\name{.get_signs_cbsft}
\alias{.get_signs_cbsft}
\title{Get signs of a circular binary segmentation fixed threshold}
\usage{
.get_signs_cbsft(obj)
}
\arguments{
\item{obj}{cbsft object}
}
\value{
a vector containing (1,-1,0)
}
\description{
The reason this is a private function is because it's somewhat non-standard.
It's unclear what "order" the jumps of a fixed threshold algorithm should
be. In this procedure, we order the jumps by their "rounds" (in 
\code{obj$tree$Get(\"active\")}), but it's unclear how the jumps are ordered
within the same round.
}
\details{
The returned vector has length equal to the number of nodes in the 
circular binary segmentation tree, and all the active nodes are in
the front of the vector. All the inactive nodes are towards the back of
the returned vector with element 0.
}
