% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_comparison_ft.R
\name{.list_comparison.cbsft}
\alias{.list_comparison.cbsft}
\title{List comparisons for circular binary segmentation fixed threshold object}
\usage{
\method{.list_comparison}{cbsft}(obj)
}
\arguments{
\item{obj}{cbsft object}
}
\value{
a list of lists of matrices
}
\description{
Returns a list with number of elements equal to the number of nodes
in \code{obj}. Each element of the list is a list with elements
\code{winning} and \code{losing}, each a matrix with 4 columns.
\code{winning} only has 1 row, denoting the comparison that was selected
at a particular step of ciruclar binary segmentation.
\code{losing} contains all the other comparisons not chosen.
}
\details{
The 4 columns represent (from left to right) the first (left) index of the left shoulder,
the first (left) index of the hump, the last (right) index of the hump,
and the last (right) index of the right shoulder. 

If an element in the returned list has \code{winning} set to 4 \code{NA}'s,
it means for that particular split, all possible future splits resulted in
a cusum statistic lower than the desired threshold.
}
