% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbs-helper.R
\name{.make_semat}
\alias{.make_semat}
\title{Matrix of selection information at this branch}
\usage{
.make_semat(m, s, e, intervals, y, thresh)
}
\arguments{
\item{m}{indices of intervals that are to be considered (i.e. are contained
in the relevant interval (s:e) at runtime of wbs()}

\item{s}{start location of the current binseg call.}

\item{e}{end location of the current binseg call.}

\item{intervals}{set of random intervals, drawn between 1 and 60}
}
\description{
Matrix of selection information at this branch
}
