% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_comparison_ft.R
\name{.remove_singleton_nodes}
\alias{.remove_singleton_nodes}
\title{Remove singleton nodes}
\usage{
.remove_singleton_nodes(node_vec)
}
\arguments{
\item{node_vec}{Vector of node names}
}
\value{
vector of node names
}
\description{
This a specialized function to remove nodes (based on their names)
that start and end with the same index. This is to prevent degenercy 
in the methods.
}
