% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_interval.R
\name{.select_index}
\alias{.select_index}
\title{Selects the index of \code{vec} that is larger than \code{alpha}. Or
something like that.}
\usage{
.select_index(vec, alpha, lower = T)
}
\description{
Selects the index of \code{vec} that is larger than \code{alpha}. Or
something like that.
}
