% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbsfs.R
\name{.split_node_cbs}
\alias{.split_node_cbs}
\title{Create the children nodes for circular binary segmentation}
\usage{
.split_node_cbs(node)
}
\arguments{
\item{node}{\code{Node} object}
}
\value{
a list
}
\description{
For a \code{node} (of class \code{Node}), where \code{node$breakpoint} is
a valid vector of breakpoints (of length 2), create up to three
children nodes according to the split. The first child, \code{left},
represents the left shoulder if it exists, \code{NA} otherwise. The
second child, \code{middle}, represents the hump. The third child,
\code{right}, represents the right shoulder if it exists, \code{NA} otherwise.
All three children are returned as a list.
}
