% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{.whichrise}
\alias{.whichrise}
\title{Does two things: Flags the result of the IC path, and locates first point of
\code{consec} rises in IC path}
\usage{
.whichrise(ic, consec = 2)
}
\arguments{
\item{ic}{numeric vector containing information criteria.}
}
\value{
First point at which ic rises.
}
\description{
Does two things: Flags the result of the IC path, and locates first point of
\code{consec} rises in IC path
}
