% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{dual1d_Dmat}
\alias{dual1d_Dmat}
\title{Function to crete 1d fused lasso regression matrix.}
\usage{
dual1d_Dmat(m)
}
\arguments{
\item{m}{data length}
}
\value{
1d Fused lasso regression matrix
}
\description{
Function to crete 1d fused lasso regression matrix.
}
