% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hausdorff.R
\name{enumerateJumps}
\alias{enumerateJumps}
\title{Compute the Jump Locations}
\usage{
enumerateJumps(vec, tol = 1e-04)
}
\arguments{
\item{vec}{a numeric vector}

\item{tol}{the numeric threshold to determine if a location is a changepoint}
}
\value{
a set of numeric integers.
}
\description{
Here, an idx i is a changepoint if i+1 != i.
}

