% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplist.R
\name{exists.cplist}
\alias{exists.cplist}
\title{gets the value corresponding to the (j,k)'th entry of cplist}
\usage{
\method{exists}{cplist}(cplist, j, k)
}
\description{
gets the value corresponding to the (j,k)'th entry of cplist
}
