% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplist.R
\name{extract.cplist}
\alias{extract.cplist}
\title{gets the value corresponding to the (j,k)'th entry of cplist}
\usage{
\method{extract}{cplist}(cplist, j, k)
}
\description{
gets the value corresponding to the (j,k)'th entry of cplist
}
