% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbs-helper.R
\name{extract_cp_from_tree}
\alias{extract_cp_from_tree}
\title{Helper function that takes in a tree (list of |semat|'s), and extracts the
changepoints}
\usage{
extract_cp_from_tree(tree, returntype = c("cp", "sign"))
}
\arguments{
\item{tree}{A list of |semat|'s. From the environment |env| you created with
\code{wbs()}, simply use \code{env$tree}.}

\item{returntype}{One of \code{c("cp","sign")}, for whether to return the
changepoint or the sign of teh changepoints.}
}
\description{
Helper function that takes in a tree (list of |semat|'s), and extracts the
changepoints
}
