% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flasso_fixedSteps.R
\name{fLasso_fixedSteps}
\alias{fLasso_fixedSteps}
\title{Fused lasso with fixed steps}
\usage{
fLasso_fixedSteps(y, numSteps, tol = 1e-07)
}
\arguments{
\item{y}{numeric vector to contain data}

\item{numSteps}{numeric of number of steps}

\item{tol}{tolerance to handle divide by zero instances}
}
\value{
a flFs object
}
\description{
y must not have duplicated values. This is to avoid
degenerate behavior of binary segmentation. The reported \code{y.fit}
is set the lambda where the (k+1)th jump is about to appear.
}
