% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binSeg_polyhedra.R
\name{form_polyhedra.bsFs}
\alias{form_polyhedra.bsFs}
\title{Generate polyhedra matrix from bsFs object}
\usage{
\method{form_polyhedra}{bsFs}(obj, ...)
}
\arguments{
\item{obj}{bsFs object}

\item{...}{not used}
}
\value{
An object of class polyhedra
}
\description{
Forms both Gamma matrix and u vector
}

