% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{form_polyhedra}
\alias{form_polyhedra}
\title{Generic function to generate polyhedra}
\usage{
form_polyhedra(obj, ...)
}
\arguments{
\item{obj}{object}

\item{...}{additional parameters}
}
\value{
An object of class polyhedra
}
\description{
Generic function to generate polyhedra
}

