% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{form_rows.intervals}
\alias{form_rows.intervals}
\title{Forms polyhedron's rows, given a maximizing b and all the qualifying indices}
\usage{
\method{form_rows}{intervals}(intervals, max.s, max.b, max.e, max.sign,
  qual.inds, n, y = NULL)
}
\arguments{
\item{y}{Defaults to NULL, but you can input y to assert that your contrasts
should be correct.}
}
\description{
Forms polyhedron's rows, given a maximizing b and all the qualifying indices
}
