% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_helpers.R
\name{fourjump_hybrid}
\alias{fourjump_hybrid}
\title{Function to generate four-jump mean with one spike and one plateau}
\usage{
fourjump_hybrid(n, lev)
}
\description{
Function to generate four-jump mean with one spike and one plateau
}
