% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_helpers.R
\name{fourjump_samesize}
\alias{fourjump_samesize}
\title{Function to generate four-jump mean /with same size jumps/.}
\usage{
fourjump_samesize(lev, n)
}
\description{
Function to generate four-jump mean /with same size jumps/.
}
