% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_helpers.R
\name{fourjump_spiky}
\alias{fourjump_spiky}
\title{Function to generate four-jump mean with narrow spikes}
\usage{
fourjump_spiky(n, lev)
}
\description{
Function to generate four-jump mean with narrow spikes
}
