% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbs-helper.R
\name{generate_intervals}
\alias{generate_intervals}
\title{Function to generate random intervals for wild binary segmentation.}
\usage{
generate_intervals(n, numIntervals, seed = NULL, start.end.list = NULL)
}
\arguments{
\item{n}{length of data.}

\item{numIntervals}{Number of intervals to draw.}

\item{seed}{seed number for random interval generation; defaults to NULL.}

\item{start.end.list}{Manual list of starts and ends. Literally an R list
with two equal length vectors, each named |start| and |end|}
}
\value{
List containing starts and ends and intervals etc.
}
\description{
Function to generate random intervals for wild binary segmentation.
}
