% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{get.closest}
\alias{get.closest}
\title{Helper to get index of closest element of \code{val} out of vector \code{allval}}
\usage{
get.closest(val, allval)
}
\description{
Helper to get index of closest element of \code{val} out of vector \code{allval}
}

