% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{get.intervals}
\alias{get.intervals}
\title{Gets \code{ni} random intervals whose intervals are sampled from \code{1:n}.}
\usage{
get.intervals(n, ni)
}
\arguments{
\item{n}{nonnegative integer}

\item{ni}{number of random intervals you want.}
}
\description{
Gets \code{ni} random intervals whose intervals are sampled from \code{1:n}.
}

