% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{get.polyhedron}
\alias{get.polyhedron}
\title{Function to collect polyhedrons given some output from the binary
segmentation function. Takes as input list contatining *list objects.}
\usage{
get.polyhedron(binseg.results, thresh, verbose = F)
}
\description{
Function to collect polyhedrons given some output from the binary
segmentation function. Takes as input list contatining *list objects.
}

