% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-helper.R
\name{get.power}
\alias{get.power}
\title{Get power from pvalues}
\usage{
get.power(pvals, loc = n/2)
}
\arguments{
\item{pvals}{list of numeric vector of p-values}

\item{loc}{location to condition on (can be a vector)}
}
\description{
Get power from pvalues
}
