% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize_new.R
\name{get_basics}
\alias{get_basics}
\title{Collecting the basic information to be shared across importance sampling
replicates.}
\usage{
get_basics(y, poly, v, sigma, bits)
}
\description{
Collecting the basic information to be shared across importance sampling
replicates.
}
