% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{get_basis}
\alias{get_basis}
\title{Get basis vector of the residual subspace between the two spanned by
\code{c(cps_old)} and \code{c(cps_old, cp_new)}.}
\usage{
get_basis(cps_old, cp_new, n)
}
\description{
Get basis vector of the residual subspace between the two spanned by
\code{c(cps_old)} and \code{c(cps_old, cp_new)}.
}
