% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_cp_from_piecewise_mean}
\alias{get_cp_from_piecewise_mean}
\title{Get changepoint locations from a piecewise constant mean.}
\usage{
get_cp_from_piecewise_mean(mn, cp)
}
\arguments{
\item{mn}{data vector.}
}
\description{
Get changepoint locations from a piecewise constant mean.
}
