% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{get_ic}
\alias{get_ic}
\title{Function that takes changepoint location vector \code{cp} (in the order that
they entered, from fixed step SBS of fixed step WBS), and calculates BIC of
each intermediate changepoint model (changepoint model = piecewise constant
underlying mean Gaussian model).}
\usage{
get_ic(
  cp,
  y,
  sigma,
  consec = 2,
  maxsteps = length(cp),
  type = "bic",
  verbose = FALSE
)
}
\arguments{
\item{cp}{Vector of changepoints}

\item{y}{Data vector.}

\item{sigma}{Noise standard deviation.}

\item{verbose}{\code{TRUE} to print progress}
}
\value{
A list containing two elements: a numeric vector of BICs, and a list
    containing basis vectors \eqn{a} of the projection \eqn{P=aa^T}from one
    changepoint column basis space to the next.
}
\description{
Function that takes changepoint location vector \code{cp} (in the order that
they entered, from fixed step SBS of fixed step WBS), and calculates BIC of
each intermediate changepoint model (changepoint model = piecewise constant
underlying mean Gaussian model).
}
