% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binSeg_fixedSteps.R
\name{get_jump_cusum.bsFs}
\alias{get_jump_cusum.bsFs}
\title{Get the cusum for jumps for bsFs objects}
\usage{
\method{get_jump_cusum}{bsFs}(obj, sorted = F, ...)
}
\arguments{
\item{obj}{bsFs object}

\item{sorted}{boolean}

\item{...}{not used}
}
\value{
vector of cusum numerics
}
\description{
Enumerates the cusum for each jump. Sorted = F will return the jumps in order
of occurance in the binSeg algorithm. Sorted = T will list the jumps
in numeric order
}

