% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_tree.R
\name{get_jump_cusum.Node}
\alias{get_jump_cusum.Node}
\title{Get the cusum for jumps for Node objects}
\usage{
\method{get_jump_cusum}{Node}(obj, sorted = F, ...)
}
\arguments{
\item{obj}{Node object}

\item{sorted}{boolean}

\item{...}{not used}
}
\value{
vector of cusum numerics
}
\description{
Enumerates the cusum for each jump. Sorted = F will return the jumps in order
of occurance in the binSeg algorithm. Sorted = T will list the jumps
in numeric order
}

