% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binSeg_fixedSteps.R
\name{get_jumps.bsFs}
\alias{get_jumps.bsFs}
\title{Get jumps from bsFs objects}
\usage{
\method{get_jumps}{bsFs}(obj, sorted = F, ...)
}
\arguments{
\item{obj}{bsFs object}

\item{sorted}{boolean}

\item{...}{not used}
}
\value{
vector of jumps
}
\description{
Enumerates the jumps. Sorted = F will return the jumps in order
of occurance in the binSeg algorithm. Sorted = T will list the jumps
in numeric order
}

