% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hausdorff.R
\name{get_jumps.numeric}
\alias{get_jumps.numeric}
\title{Compute the Jump Locations}
\usage{
\method{get_jumps}{numeric}(obj, tol = 1e-10, ...)
}
\arguments{
\item{obj}{a numeric vector}

\item{tol}{the numeric threshold to determine if a location is a changepoint}

\item{...}{not used}
}
\value{
a set of numeric integers.
}
\description{
Here, an idx i is a changepoint if i+1 != i.
}

