% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplist.R
\name{get_last_row_val.cplist}
\alias{get_last_row_val.cplist}
\title{Extracts /last/ element of cplist. Mainly used in wbs-FS-polyhedra.}
\usage{
\method{get_last_row_val}{cplist}(cplist)
}
\arguments{
\item{cplist}{A cplist object.}
}
\value{
"val" value of the last row of cplist.
}
\description{
Extracts /last/ element of cplist. Mainly used in wbs-FS-polyhedra.
}
