% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_piecewise_mean}
\alias{get_piecewise_mean}
\title{Return piecewise mean.}
\usage{
get_piecewise_mean(y, cp)
}
\arguments{
\item{y}{data vector.}

\item{cp}{changepoint vector. Assumed to be sorted.}
}
\description{
Return piecewise mean.
}
