% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-helper.R
\name{get_sigma}
\alias{get_sigma}
\title{Gets sigma by fitting basic wbs::sbs on it, with default settings for
complexity.}
\usage{
get_sigma(y)
}
\arguments{
\item{y}{data vector}
}
\description{
Gets sigma by fitting basic wbs::sbs on it, with default settings for
complexity.
}
