% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{getcusums}
\alias{getcusums}
\title{Get all cusums, given start point \code{s} and end point \code{e}}
\usage{
getcusums(s, e, y, unsigned = FALSE, simplereturn = FALSE)
}
\arguments{
\item{s}{Starting index, between \code{1} and \code{n}}

\item{e}{Ending index, between \code{1} and \code{n}}

\item{y}{\code{n}-lengthed data vector.}

\item{unsigned}{\code{TRUE} to return alsolute value.}

\item{simplereturn}{\code{TRUE} just to return the cusums.}
}
\value{
list of information about the cusum calculations and maximizers in
    this interval.
}
\description{
Get all cusums, given start point \code{s} and end point \code{e}
}
