% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{haarbasis}
\alias{haarbasis}
\title{Calculates the unbalanced Haar basis for subvector of \code{y} or the vector
\code{a} in linear inequalities a*y>=0}
\usage{
haarbasis(s, b, e, n, y, type = c("basis", "ineq"))
}
\arguments{
\item{type}{\code{basis} to return the unbalanced haar basis, and \code{ineq}
to return linear inequality vector.}
}
\description{
Calculates the unbalanced Haar basis for subvector of \code{y} or the vector
\code{a} in linear inequalities a*y>=0
}
\examples{
y = c(rnorm(10,0,1), rnorm(10,4,1))

# Calculate Haar basis, s:e vs (b+1):e
##mybasis = haarbasis(s = 0, b = 10, e = 20, n = 20, y = y, type = "basis")

# Calculate linear inequality vectors
##myineqs = haarbasis(s = 0, b = 10, e = 20, n = 20, y = y, type = "ineq")

}

