% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs-poly.R
\name{halfspaces}
\alias{halfspaces}
\title{Calculates the halfspace vectors for the maximiiming breakpoint and all the
signs, for fixed-threshold SBS.}
\usage{
halfspaces(s, b, e, z, thresh, n, y, is.terminal.node = F, verbose = F)
}
\arguments{
\item{b}{breakpoint location.}

\item{is.terminal.node}{TRUE if no CUSUMs from this node pass the threshold.}
}
\value{
A list of two objects \code{V} and \code{u}, for the halfspaces in
    represented as V'y>u
}
\description{
Calculates the halfspace vectors for the maximiiming breakpoint and all the
signs, for fixed-threshold SBS.
}
