% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-helper.R
\name{has.loc}
\alias{has.loc}
\title{See if the vector of p-values (whose rownames are locations) \code{pv} has
location \code{loc}.}
\usage{
has.loc(pv, loc)
}
\arguments{
\item{pv}{numeric vector of p-values}

\item{loc}{location}
}
\description{
See if the vector of p-values (whose rownames are locations) \code{pv} has
location \code{loc}.
}
