% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{hausdorff}
\alias{hausdorff}
\title{Computes the Hausdorff Distance between Two Sets}
\usage{
hausdorff(set1, set2, one.sided = F)
}
\arguments{
\item{set1}{a numeric vector}

\item{set2}{a numeric vector}

\item{one.sided}{a logical to determine if a one-sided distance or 
two-sided distance is computed.}
}
\value{
a numeric (length 1)
}
\description{
The two sets, set1 and set2, must be numeric vectors. If one.sided is true,
the hausdorff returns max_(a in set1) min_(b in set2) |a-b|.
}
\details{
If either sets are empty, NA is returned.
}
