% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-helper.R
\name{i.covers.cp}
\alias{i.covers.cp}
\title{Helper to inquire about changepoint being tested, \code{cp}, being contained
in the interval set $i$.}
\usage{
i.covers.cp(i, cp)
}
\arguments{
\item{i}{changepoint set}

\item{cp}{single location}
}
\value{
TRUE if \code{cp} is covered by \code{i}
}
\description{
Helper to inquire about changepoint being tested, \code{cp}, being contained
in the interval set $i$.
}
