% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{ic_flag}
\alias{ic_flag}
\title{Helper to identify and flag the outcome of an IC vector.}
\usage{
ic_flag(ic, consec = 2, cp, maxsteps)
}
\description{
Helper to identify and flag the outcome of an IC vector.
}
