% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{ic_to_poly}
\alias{ic_to_poly}
\title{Takes IC vector and creates polyhedra.}
\usage{
ic_to_poly(obj)
}
\arguments{
\item{obj}{object of class \code{ic} from \code{get_ic()}, which contains
things needed for sequential information criteria (ic) comparison.}

\item{resid}{list of residual vectors}
}
\value{
Object of class \code{polyhedra}, for sequential IC comparisons.
}
\description{
Takes IC vector and creates polyhedra.
}
