% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{ic_wrapper}
\alias{ic_wrapper}
\title{Embed stopping time into the path.}
\usage{
ic_wrapper(obj, y, consec = 2, maxsteps = length(obj$cp), sigma,
  type = "bic")
}
\arguments{
\item{obj}{A list object containing an element named \code{"cp"}, which is a
integer vector with changepoints, in the order they were selected by the
sequential algorithm}

\item{consec}{Number of rises in IC before stop is to happen. Defaults to 2.}

\item{sigma}{Standard deviation of noise generating the data, in \code{y}}

\item{type}{One of \code{c("bic","ebic","aic")}. Defaults to
\code{"bic"}. Only BIC is possible for now.}
}
\value{
list containing two objects: ''poly'' and ''stoptime''.
}
\description{
Embed stopping time into the path.
}
