% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{inference_bsfs}
\alias{inference_bsfs}
\title{Conducts post-selection inference procedure for additive noise binary
segmentation inference -- segment test inference for all detected
changepoints -- for a given n-lengthed data vector y, with or without IC
stopping.}
\usage{
inference_bsfs(y = y, max.numSteps = 20, consec = 2, sigma,
  icstop = TRUE, postprocess = TRUE, locs = 1:length(y),
  numIS = 100, sigma.add = 0.2, bits = 50,
  inference.type = c("rows", "pre-multiply"), numIntervals = length(y),
  max.numIS = 2000, verbose = FALSE, min.num.things = 10,
  added.noise = NULL, mc.cores = 1, start.time = NULL,
  how.close = 5, mn = NULL, sim.options = list(retain.only.null.cases
  = FALSE), bootstrap.inds = NULL)
}
\arguments{
\item{y}{data vector}

\item{max.numSteps}{The (maximum) number of algorithm steps you'd like to
do, prior to BIC. If you say \code{icstop=FALSE}, then this is the
actual, fixed number of algorithm steps used.}

\item{consec}{The number of rises you'd like as a stopping rule, in terms of
BIC.}

\item{icstop}{Whether to use IC stopping. Defaults to \code{TRUE}.}

\item{postprocess}{\code{TRUE} if you'd like to centroid cluster the
detected changepoints.}

\item{locs}{Which locations to test. Defaults to \code{1:length(y)}.}

\item{verbose}{Whether to print progress of}

\item{added.noise}{is a manually inputted additive noise vector. This must
be generated from i.i.d. Gaussian noise with \code{sigma} standard
deviation. A rough check is in place, but really just trusting the user
at this point.}

\item{mc.cores}{Whether to run importance sampler on multiple cores. Not
recommended for normal user.}

\item{start.time}{Optional argument, for timing things when
\code{verbose=TRUE}}

\item{how.close}{How close would you like points to be in each cluster, for
the centroid clustering, to be.}

\item{return}{An object of class \code{bsfs} with p-values.}
}
\description{
Conducts post-selection inference procedure for additive noise binary
segmentation inference -- segment test inference for all detected
changepoints -- for a given n-lengthed data vector y, with or without IC
stopping.
}
