% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{inference_wbs}
\alias{inference_wbs}
\title{Does a single randomized wbs (rwbs) inference for a given y}
\usage{
inference_wbs(
  y = y,
  max.numSteps = 20,
  numIntervals = length(y),
  intervals = NULL,
  consec = 2,
  sigma,
  postprocess = TRUE,
  how.close = 5,
  better.segment = FALSE,
  locs = 1:length(y),
  numIS = 100,
  inference.type = inference.type,
  bits = 1000,
  max.numIS = 2000,
  verbose = FALSE,
  mc.cores = 1,
  min.num.things = 30
)
}
\arguments{
\item{numIntervals}{Defaults to \code{length(y)}, or if |intervals| is
provided, the length of that.}

\item{intervals}{An object of class |intervals|.}

\item{postprocess}{\code{TRUE} if you'd like to centroid cluster the
detected changepoints.}

\item{howclose}{How close would you like points to be in each cluster, for
the centroid clustering, to be.}
}
\value{
List of simulation results (p-values, locations tested, etc.)
}
\description{
Does a single randomized wbs (rwbs) inference for a given y
}
