% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{intervals}
\alias{intervals}
\title{Constructor for intervals object.}
\usage{
intervals(
  numIntervals,
  n,
  comprehensive = FALSE,
  existing = NULL,
  distance = 0,
  maxlength = n - 1,
  precuts = NULL,
  draw.fac = 5
)
}
\arguments{
\item{existing}{A 2-row numeric matrix containing start and end points.
Column names should be names "s" and "e" respectively.}

\item{distance}{minimum distance between s and e, i.e. e is at least
s+distance.}

\item{maxlength}{maximum distance between s and e, i.e. (e-s <= maxlength)}

\item{precuts}{precut locations. Defaults to NULL.}

\item{nrow}{Number of rows in the empty matrix}
}
\value{
creates an all-NA matrix of dimension nrow x 3. The first two
    columns must be the numeric (no check yet), but the last column can be
    of any type you want. Initializes to numeric.
}
\description{
Constructor for intervals object.
}
