% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbsfs.R
\name{is.na.Node}
\alias{is.na.Node}
\title{Placeholder function}
\usage{
\method{is.na}{Node}(x)
}
\arguments{
\item{x}{\code{Node} object}
}
\value{
FALSE
}
\description{
Included for technical coding reasons.
}
