% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binSeg_fixedSteps.R
\name{is_valid.bsFs}
\alias{is_valid.bsFs}
\title{is_valid for bsFs}
\usage{
\method{is_valid}{bsFs}(obj)
}
\arguments{
\item{obj}{bsFs object}
}
\value{
TRUE if valid
}
\description{
is_valid for bsFs
}
