% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsft.R
\name{is_valid.bsft}
\alias{is_valid.bsft}
\title{Checks of object is of class "bsft", produced by binSeg_fixedThresh().}
\usage{
\method{is_valid}{bsft}(obj)
}
\arguments{
\item{obj}{bsft object}
}
\description{
Checks of object is of class "bsft", produced by binSeg_fixedThresh().
}
