% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{is_valid.intervals}
\alias{is_valid.intervals}
\title{Method to check if object is a valid object of the |interval| class.}
\usage{
\method{is_valid}{intervals}(obj)
}
\description{
Method to check if object is a valid object of the |interval| class.
}
