% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyhedra.R
\name{is_valid.polyhedra}
\alias{is_valid.polyhedra}
\title{Check if polyhedra is valid}
\usage{
\method{is_valid}{polyhedra}(obj)
}
\arguments{
\item{obj}{object of class polyhedra}
}
\value{
TRUE if valid
}
\description{
Check if polyhedra is valid
}
