% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{is_valid}
\alias{is_valid}
\title{Checks S3 object validity}
\usage{
is_valid(obj)
}
\arguments{
\item{obj}{The object to check}
}
\value{
void
}
\description{
Generic function that fails noisily with a stop message if the object is
invalid. Otherwise, nothing happens.
}
