% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildBinSeg_fixedStep.R
\name{is_valid.wbsFs}
\alias{is_valid.wbsFs}
\title{is_valid for wbs}
\usage{
\method{is_valid}{wbsFs}(obj)
}
\arguments{
\item{obj}{wbs object}
}
\value{
TRUE if valid
}
\description{
is_valid for wbs
}
