% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_tree.R
\name{isValid.Node}
\alias{isValid.Node}
\title{Check whether tree is valid or not}
\usage{
\method{isValid}{Node}(obj)
}
\arguments{
\item{obj}{The tree (of class Node)}
}
\value{
TRUE if valid
}
\description{
Check whether tree is valid or not
}

