% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyhedra.R
\name{isValid.polyhedra}
\alias{isValid.polyhedra}
\title{Check if polyhedra is valid}
\usage{
\method{isValid}{polyhedra}(obj)
}
\arguments{
\item{obj}{object of class polyhedra}
}
\value{
TRUE if valid
}
\description{
Check if polyhedra is valid
}

