% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsfs.R
\name{jump_cusum.bsfs}
\alias{jump_cusum.bsfs}
\title{Get the cusum for jumps for bsfs objects}
\usage{
\method{jump_cusum}{bsfs}(obj, sorted = F, ...)
}
\arguments{
\item{obj}{bsfs object}

\item{sorted}{boolean}

\item{...}{not use}
}
\value{
vector of cusum numerics
}
\description{
Enumerates the cusum for each jump. Sorted = F will return the jumps in order
of occurance in the binSeg algorithm. Sorted = T will list the jumps
in numeric order
}
