% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{jump_cusum}
\alias{jump_cusum}
\title{Generic function to count the number of jump cusum}
\usage{
jump_cusum(obj, ...)
}
\arguments{
\item{obj}{object}

\item{...}{additional parameters}
}
\value{
vector of cusums
}
\description{
Generic function to count the number of jump cusum
}
