% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flasso_fixedSteps.R
\name{jump_lambda.flFs}
\alias{jump_lambda.flFs}
\title{Get lambdas for flFs objects}
\usage{
\method{jump_lambda}{flFs}(obj, ...)
}
\arguments{
\item{obj}{flFs object}

\item{...}{not used}
}
\value{
vector of decreasing lambdas
}
\description{
Enumerates the decreasing sequence of lambdas for the jumps.
Sorted = F will return the jumps in order
of occurance in the binSeg algorithm. Sorted = T will list the jumps
in numeric order
}

