% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{jump_lambda}
\alias{jump_lambda}
\title{Generic function to count the number of jump lambdas}
\usage{
jump_lambda(obj, ...)
}
\arguments{
\item{obj}{object}

\item{...}{additional parameters}
}
\value{
vector of lambdas
}
\description{
Generic function to count the number of jump lambdas
}
