% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{jump_sign}
\alias{jump_sign}
\title{Generic function to get a matrix of jumps and signs}
\usage{
jump_sign(obj, sorted = F)
}
\arguments{
\item{obj}{object}

\item{sorted}{boolean}
}
\value{
2-column matrix of jumps and signs
}
\description{
Generic function to get a matrix of jumps and signs
}
