% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsfs.R
\name{jumps.bsfs}
\alias{jumps.bsfs}
\title{Get jumps from bsfs objects}
\usage{
\method{jumps}{bsfs}(obj, sorted = F, ...)
}
\arguments{
\item{obj}{bsfs object}

\item{sorted}{boolean}

\item{...}{not used}
}
\value{
vector of jumps
}
\description{
Enumerates the jumps. Sorted = F will return the jumps in order
of occurance in the binSeg algorithm. Sorted = T will list the jumps
in numeric order
}
