% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbsfs.R
\name{jumps.cbsfs}
\alias{jumps.cbsfs}
\title{Get jumps from cbsfs objects}
\usage{
\method{jumps}{cbsfs}(obj, sorted = T, unique = TRUE, cbs = FALSE,
  remove.na = FALSE, ...)
}
\arguments{
\item{obj}{cbs object}

\item{sorted}{boolean}

\item{unique}{prunes the changepoints to be unique. Not to be used for CBS}

\item{...}{not used}
}
\value{
vector of all the jump locations
}
\description{
Enumerates the jumps for circular binary segmentation. If \code{sorted}
is true, then only the unique changepoints are reported, and no
\code{NA} is recorded. If \code{sorted} is false, then each pair of
elements in the result denote the jump locations, and \code{NA} means
the left (first) element of the hump is the first index.
}
