% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbsft.R
\name{jumps.cbsft}
\alias{jumps.cbsft}
\title{Get jumps from cbsft objects}
\usage{
\method{jumps}{cbsft}(obj, ...)
}
\arguments{
\item{obj}{cbs object}

\item{...}{not used}
}
\value{
vector of all the jump locations
}
\description{
Enumerates the jumps for circular binary segmentation.
}
