% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flasso_fixedSteps.R
\name{jumps.flFs}
\alias{jumps.flFs}
\title{Get jumps from flFs objects}
\usage{
\method{jumps}{flFs}(obj, sorted = F, ...)
}
\arguments{
\item{obj}{flFs object}

\item{sorted}{boolean}

\item{...}{not used}
}
\value{
vector of jumps
}
\description{
Enumerates the jumps. Sorted = F will return the jumps in order
of occurance in the binSeg algorithm. Sorted = T will list the jumps
in numeric order
}
