% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_tree.R
\name{jumps.Node}
\alias{jumps.Node}
\title{Get jumps from a tree}
\usage{
\method{jumps}{Node}(obj, sorted = F, ...)
}
\arguments{
\item{obj}{object of class Node}

\item{sorted}{boolean}

\item{...}{not used}
}
\value{
vector of jumps
}
\description{
Enumerates the jumps. Sorted = F will return the jumps in order
of occurance in the binSeg algorithm. Sorted = T will list the jumps
in numeric order
}
