% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{jumps}
\alias{jumps}
\title{Generic function to count the number of jumps}
\usage{
jumps(obj, ...)
}
\arguments{
\item{obj}{object}

\item{...}{additional parameters}
}
\value{
vector of jumps
}
\description{
Generic function to count the number of jumps
}
