% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.R
\name{make.qqplot.background}
\alias{make.qqplot.background}
\title{makes QQ plot background without any lines.}
\usage{
make.qqplot.background(plot.title = "")
}
\description{
makes QQ plot background without any lines.
}

