% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.R
\name{make.qqplot.legend}
\alias{make.qqplot.legend}
\title{makes QQ plot legend on existing plot.}
\usage{
make.qqplot.legend(deltas, pch = 16, pcols = rep("red", length(deltas)),
  where.legend = "bottomright")
}
\description{
makes QQ plot legend on existing plot.
}

