% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{make.tDb}
\alias{make.tDb}
\title{Helper function to make basis vectors for the piecewise constant space
broken at changepoints at |cps|. |tDb| just stands for D^T[,-b].}
\usage{
make.tDb(cps = c(), n)
}
\description{
Helper function to make basis vectors for the piecewise constant space
broken at changepoints at |cps|. |tDb| just stands for D^T[,-b].
}
