% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{make.v.fixed.thresh}
\alias{make.v.fixed.thresh}
\title{Function to obtain contrast, given the output of binseg(), the SBS algorithm
with fixed threshold.}
\usage{
make.v.fixed.thresh(test.b, bs.output)
}
\arguments{
\item{test.b}{is the location that we want to test.}

\item{bs.output}{list that contains G,u,y,blist,zlist. Manually bundled by
user.}
}
\description{
Function to obtain contrast, given the output of binseg(), the SBS algorithm
with fixed threshold.
}

