% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{make.v}
\alias{make.v}
\title{Function to obtain contrast, given the output of binseg(), the SBS algorithm
with fixed threshold.}
\usage{
make.v(test.b, B, Z, n)
}
\arguments{
\item{test.b}{is the location that we want to test.}

\item{B}{vector of breakpoints to be considered}

\item{Z}{signs of B, as}

\item{n}{length of contrast}
}
\description{
Function to obtain contrast, given the output of binseg(), the SBS algorithm
with fixed threshold.
}

