% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{make.v.simple}
\alias{make.v.simple}
\title{Simpler version of make.v
with fixed threshold.}
\usage{
make.v.simple(b, s, e, n, dir)
}
\arguments{
\item{b}{is the location that we want to test.}
}
\description{
Simpler version of make.v
with fixed threshold.
}

