% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{make_all_segment_contrasts_from_cp}
\alias{make_all_segment_contrasts_from_cp}
\title{Take |cp| and |cp.sign| and make all segment contrast vectors.}
\usage{
make_all_segment_contrasts_from_cp(cp, cp.sign, n,
  scaletype = c("segmentmean", "unitnorm"))
}
\arguments{
\item{cp}{An integer vector of changepoints.}

\item{cp.sign}{An integer vector of changepoint signs.}

\item{n}{Data length.}
}
\value{
A list of test contrast vectors, whose names are the signed
    changepoint locations.
}
\description{
Take |cp| and |cp.sign| and make all segment contrast vectors.
}
