% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{make_all_segment_contrasts_from_wbs}
\alias{make_all_segment_contrasts_from_wbs}
\title{Makes segment contrasts with endpoints as the winning intervals' start/ends}
\usage{
make_all_segment_contrasts_from_wbs(
  wbsfs_obj,
  cps = NULL,
  scaletype = c("segmentmean", "unitnorm")
)
}
\arguments{
\item{wbsfs_obj}{An object of class \code{wbsfs}.}

\item{cps}{if not \code{NULL}, only form contrasts from changepoints in this
set.}
}
\value{
A list of test contrast vectors, whose names are the signed
    changepoint locations.
}
\description{
Makes segment contrasts with endpoints as the winning intervals' start/ends
}
