% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbs-helper.R
\name{make_contrast}
\alias{make_contrast}
\title{Helper function to /manually/ make contrasts.}
\usage{
make_contrast(test.bp, adj.bps, sn, n)
}
\arguments{
\item{adj.bps}{Directly adjacent breakpoint locations.}

\item{sn}{Sign (direction) of proposed breakpoint at |test.bps|; use +1 or -1.}

\item{n}{length of data.}

\item{test.bps}{Breakpoint location to test.}
}
\description{
Helper function to /manually/ make contrasts.
}
\examples{
make_contrast(20,c(1,40),+1,60)
}
