% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{make_pw_mean}
\alias{make_pw_mean}
\title{Make n-length vector that induces piecewise mean that is cut at cp. Used for
simulation examples.}
\usage{
make_pw_mean(y, cp = c())
}
\arguments{
\item{y}{Original data vector.}

\item{cp}{Changepoint location. Defaults to \code{c()}.}
}
\value{
n-length vector with means
}
\description{
Make n-length vector that induces piecewise mean that is cut at cp. Used for
simulation examples.
}
