% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{make_segment_inds}
\alias{make_segment_inds}
\title{Helper to make segment indices.}
\usage{
make_segment_inds(cp, n)
}
\arguments{
\item{cp}{Changepoints.}

\item{n}{Data length.}
}
\description{
Helper to make segment indices.
}
