% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbsft-poly.R
\name{max_of_max_obj}
\alias{max_of_max_obj}
\title{Creates halfspaces for the main maximize-max-cusums.}
\usage{
max_of_max_obj(obj)
}
\arguments{
\item{obj}{Result of running WBS.}
}
\description{
Creates halfspaces for the main maximize-max-cusums.
}
