% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbsft-poly.R
\name{max_of_max_semat}
\alias{max_of_max_semat}
\title{Creates halfspaces for the main maximize-max-cusums.  TODO: many of the rows
overlap; keep track of this and don't add duplicates. Currently, this is not
done carefully here; but rather, during aggregation of the polyhedra, the
duplicates are eliminated.}
\usage{
max_of_max_semat(semat, obj)
}
\arguments{
\item{semat}{A matrix with results from a node of an obj$tree.}
}
\description{
Creates halfspaces for the main maximize-max-cusums.  TODO: many of the rows
overlap; keep track of this and don't add duplicates. Currently, this is not
done carefully here; but rather, during aggregation of the polyhedra, the
duplicates are eliminated.
}
