% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{maximize.intervals}
\alias{maximize.intervals}
\title{Get the cusum maximizer, among qualifying guys.}
\usage{
\method{maximize}{intervals}(intervals, qual.inds)
}
\description{
Get the cusum maximizer, among qualifying guys.
}
