% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{Mclapply}
\alias{Mclapply}
\title{Experimental: a convenience wrapper to \code{mclapply()} that takes
\code{start.time}.}
\usage{
Mclapply(nsim, myfun, mc.cores, start.time = NULL)
}
\description{
Experimental: a convenience wrapper to \code{mclapply()} that takes
\code{start.time}.
}
