% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue.R
\name{mpfr.tnorm.surv}
\alias{mpfr.tnorm.surv}
\title{Temporarily added from selectiveInference package.
Returns Prob(Z>z | Z in [a,b]), where mean can be a vector, using
multi precision floating point calculations thanks to the Rmpfr package}
\usage{
mpfr.tnorm.surv(z, mean = 0, sd = 1, a, b, bits = NULL)
}
\description{
Temporarily added from selectiveInference package.
Returns Prob(Z>z | Z in [a,b]), where mean can be a vector, using
multi precision floating point calculations thanks to the Rmpfr package
}
