% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_interval.R
\name{my.one.sided.ci}
\alias{my.one.sided.ci}
\title{A more barebones, readable version of a one-sided CI from a
right-sided-alternative TG test. Matches the output of
\code{confidence_interval()} when using the \code{alternative="one-sided"}
option.}
\usage{
my.one.sided.ci(
  y,
  poly,
  contrast,
  sigma,
  alpha,
  fac = 10,
  gridsize = 250,
  griddepth = 2
)
}
\description{
A more barebones, readable version of a one-sided CI from a
right-sided-alternative TG test. Matches the output of
\code{confidence_interval()} when using the \code{alternative="one-sided"}
option.
}
