% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-helper.R
\name{onejump_data}
\alias{onejump_data}
\title{Toy data generating function, of upward one-jump signal, for use in Rmd
 simulation files.}
\usage{
onejump_data(seed = NULL, n, lev, sigma, type = c("data", "mean"))
}
\arguments{
\item{seed}{Seed to generate data from}

\item{n}{Data length}

\item{lev}{jump size}

\item{sigma}{Data noise}

\item{type}{data}
}
\description{
Toy data generating function, of upward one-jump signal, for use in Rmd
 simulation files.
}
