% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{partition_TG}
\alias{partition_TG}
\title{From polyhedron and data vector and contrast, gets the probability that vtY
is in the polyhedron, conditional on PvperpY. i.e. the probability that Y
stays in the polyhedron, fixing n-1 dimensions of it.}
\usage{
partition_TG(y, poly, v, sigma, nullcontrast = 0, bits = 50, reduce,
  correct.ends = FALSE, shift = NULL, ic.poly = NULL, warn = TRUE)
}
\arguments{
\item{y}{data}

\item{poly}{polyhedra object produced form \code{polyhedra(wbs_object)}}

\item{v}{contrast vector}

\item{sigma}{data noise (standard deviation)}

\item{nullcontrast}{the null value of \eqn{v^T\mu}, for \eqn{\mu = E(y)}.}
}
\value{
list of two vectors: denominators and numerators, each named
    \code{denom} and \code{numer}.
}
\description{
From polyhedron and data vector and contrast, gets the probability that vtY
is in the polyhedron, conditional on PvperpY. i.e. the probability that Y
stays in the polyhedron, fixing n-1 dimensions of it.
}
