% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize_new.R
\name{partition_TG2}
\alias{partition_TG2}
\title{Reusing the basic information that is shared across all importance sampling
replicates.}
\usage{
partition_TG2(obj.basic, Gw, bits = 2000)
}
\description{
Reusing the basic information that is shared across all importance sampling
replicates.
}
