% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{perm.t.test}
\alias{perm.t.test}
\title{Conducts a permutation t-test, given two subvectors
Example doesn't work now, but here it is: EXAMPLES/perm.t.test.example.R}
\usage{
perm.t.test(vec1, vec2, nsim = 1000)
}
\description{
Conducts a permutation t-test, given two subvectors
Example doesn't work now, but here it is: EXAMPLES/perm.t.test.example.R
}

