% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{piecewise_mean}
\alias{piecewise_mean}
\title{Gets piecewise mean, given changepoint locations.}
\usage{
piecewise_mean(y, cp)
}
\arguments{
\item{y}{Data.}

\item{cp}{Changepoints.}
}
\value{
Piecewise constant vector of data whose entries are sample means of
    segments.
}
\description{
Gets piecewise mean, given changepoint locations.
}
