% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.CpVector}
\alias{plot.CpVector}
\title{Plot function for CpVector}
\usage{
\method{plot}{CpVector}(x, col.dots = defaultColors(-1, 0.75),
  col.line = defaultColors(1), ylim = NA, ...)
}
\arguments{
\item{x}{the CpVector instance}

\item{col.dots}{the color of the data points (x$data)}

\item{col.line}{the color of the true signal (x$jump.height and x$jump.idx)}

\item{ylim}{the graphical parameter ylim}

\item{...}{additional graphical parameters such as pch, par, etc.}
}
\value{
void
}
\description{
Plot function for CpVector
}
\examples{
set.seed(10)
res <- CpVector(n = 100, jump.height = 1:4, jump.loc = c(.25,.5,.75))
defaultPlotDefaults()
plot(res, ylab = "Value", xlab = "n")
}
