% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{plot.intervals}
\alias{plot.intervals}
\title{After making intervals, you can attempt to plot them.}
\usage{
\method{plot}{intervals}(obj)
}
\description{
After making intervals, you can attempt to plot them.
}
