% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{plot.v}
\alias{plot.v}
\title{Takes a contrast vector v and optionally B & Z, and plots it}
\usage{
\method{plot}{v}(v, B = NULL, Z = NULL)
}
\arguments{
\item{v}{Numeric vector containing contrast vector.}
}
\description{
Takes a contrast vector v and optionally B & Z, and plots it
}

