% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue.R
\name{poly.pval}
\alias{poly.pval}
\title{Temporarily added from selectiveInference package.}
\usage{
poly.pval(y, G, u, v, sigma, bits = NULL)
}
\description{
Temporarily added from selectiveInference package.
}
