% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue.R
\name{poly.pval2}
\alias{poly.pval2}
\title{Takes vup, vlo and v and returns list(pv,vlo,vup). Originally from the
selectiveInfernece package. Modified to take a polyhedra class object
\code{poly}.}
\usage{
poly.pval2(y, poly = NULL, v, sigma, vup = NULL, vlo = NULL,
  bits = NULL, reduce = FALSE, correct.ends = FALSE)
}
\arguments{
\item{y}{data vector}

\item{v}{Contrast vector.}

\item{sigma}{noise level}

\item{vup}{vup}

\item{vlo}{vlo}

\item{bits}{Number of decimal points for higher precision calculation of
pvalue. (i.e. precision of Rmpfr)}
}
\value{
List of vup, vlo and pv.
}
\description{
Takes vup, vlo and v and returns list(pv,vlo,vup). Originally from the
selectiveInfernece package. Modified to take a polyhedra class object
\code{poly}.
}
