% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildBinSeg_fixedStep_polyhedra.R
\name{poly_from_snapshot}
\alias{poly_from_snapshot}
\title{Takes in object and step, and collects the selection event at that step!
This means the characterization of the cusum statistic at (b.max, m.max and
s.max) being the largest among all the competing intervals at that step.}
\usage{
poly_from_snapshot(obj, mystep, reduce = FALSE, latest.poly = NULL)
}
\arguments{
\item{obj}{\code{wbsFt} object.}

\item{mystep}{step in the path}

\item{reduce}{If TRUE, then does a Vup/Vlo comparison to see if you
should add (chunks) of r}

\item{latest.poly}{If \code{reduce} is set to TRUE, this is the latest
polyhedron outside of. This is necessary because polyhedron
approximations only work if the checking for change in (Vup, Vlo) is
done /cumulatively/. Defaults to NULL.}
}
\value{
a polyhedra object with the selection event at that step.
}
\description{
Takes in object and step, and collects the selection event at that step!
This means the characterization of the cusum statistic at (b.max, m.max and
s.max) being the largest among all the competing intervals at that step.
}
