% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue.R
\name{poly_pval2_from_vlist}
\alias{poly_pval2_from_vlist}
\title{If a list of contrast vectors are supplied, use this.}
\usage{
poly_pval2_from_vlist(y, poly, vlist, sigma, bits = 5000)
}
\description{
If a list of contrast vectors are supplied, use this.
}
