% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue.R
\name{poly_pval_bootsub_inner}
\alias{poly_pval_bootsub_inner}
\title{Calculating TG p-value from bootstrapping centered residuals}
\usage{
poly_pval_bootsub_inner(
  Vlo,
  Vup,
  vty,
  v,
  y = NULL,
  nboot = 1000,
  bootmat = NULL,
  weight = FALSE,
  bootmat.times.v = NULL,
  adjustmean,
  pad = FALSE
)
}
\arguments{
\item{y}{data vector.}

\item{bootmat}{vector whose rows are bootstrapped y's.}

\item{weight}{if TRUE, returns denominator of TG; if FALSE, returns entire
p-value.}

\item{bootmat.times.v}{vector of bootmat times y.}
}
\value{
p-value or weight, depending on \code{weight} input.
}
\description{
Calculating TG p-value from bootstrapping centered residuals
}
