% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue.R
\name{poly_pval_bootsub_large}
\alias{poly_pval_bootsub_large}
\title{For large |nboot|, calculating TG p-value from bootstrapped residuals by
chunking into 10000 replicates each. TODO In the future, we want to do /not/
repeat the bootstrap for each v, but do it in batch.}
\usage{
poly_pval_bootsub_large(y, G, v, nboot.max = 100 * 10000, sigma,
  adjustmean = mean(y), pad = FALSE, vlist = NULL, ridlong = FALSE,
  stable.thresh = 0.01)
}
\arguments{
\item{y}{data vector.}

\item{G}{Polyhedron gamma matrix.}

\item{v}{contrast vector}

\item{nboot.max}{The number of bootstraps replicates in total.}

\item{vlist}{A list of contrast vectors. If supplied, then v will be
ignored.}
}
\description{
For large |nboot|, calculating TG p-value from bootstrapped residuals by
chunking into 10000 replicates each. TODO In the future, we want to do /not/
repeat the bootstrap for each v, but do it in batch.
}
