% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue.R
\name{poly_pval_from_inner_products}
\alias{poly_pval_from_inner_products}
\title{Computes TG p-value and related objects (vlo,vup,vty,denom,numer) in the
case that Gy and Gv are pre-calculated; the option
inference.type="pre-multiply".}
\usage{
poly_pval_from_inner_products(Gy, Gv, v, y, sigma, u, bits = 1000,
  warn = TRUE)
}
\arguments{
\item{Gy}{G times y}

\item{Gv}{G times v}

\item{v}{constrast vector}

\item{y}{data}

\item{sigma}{noise level}

\item{bits}{number of precision bits used for calculation of Gaussian
tails. Roughly 3*bits is the /number/ of digits that is being used.}
}
\value{
list containing pv,vlo,vup,vty,denom,numer.
}
\description{
Computes TG p-value and related objects (vlo,vup,vty,denom,numer) in the
case that Gy and Gv are pre-calculated; the option
inference.type="pre-multiply".
}
