% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue.R
\name{poly_pval_premultiply}
\alias{poly_pval_premultiply}
\title{Wrapper to *only* do the premultiply option for plain TG inference..}
\usage{
poly_pval_premultiply(y = y, v = v, obj = obj, sigma = sigma)
}
\arguments{
\item{y}{data vector}

\item{v}{contrast vector}

\item{obj}{output from running an algorithm *before* adding polyhedra information.}
}
\description{
Wrapper to *only* do the premultiply option for plain TG inference..
}
