% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs-poly.R
\name{polyhedra.bsfs}
\alias{polyhedra.bsfs}
\title{Generate polyhedra matrix from bsfs object}
\usage{
\method{polyhedra}{bsfs}(
  obj,
  numSteps = NA,
  inference.type = c("rows", "pre-multiply"),
  new.noise = NULL,
  v = NULL,
  icpoly = NULL,
  y = NULL,
  excessive = FALSE,
  record.nrows = TRUE,
  ...
)
}
\arguments{
\item{obj}{bsfs object}

\item{inference.type}{type of return -- either the actual polyhedron, or the
pre-multiplied quantities Gv, Gy and Gw.}

\item{new.noise}{New additive noise, drawn in order for randomization.}

\item{icpoly}{\code{polyhedra} class object for IC stopping. Defaults to
\code{NULL}.}

\item{excessive}{Experimental feature for adding excessive conditioning for
CUSUM signs.}

\item{...}{not used}

\item{record.rownums}{If \code{TRUE}, record the number of rows corresponding
to each algorithm step in \code{1:obj$numSteps} (or \code{1:numSteps} if
\code{numSteps} is provided. Only relevant when
\code{inference.type=='rows'}.}
}
\value{
An object of class polyhedra. Includes \code{nrow.by.step} object,
    which is the number of halfspaces (rows of gamma matrix) corresponding to
    each step, when \code{inference.type=="rows"}.
}
\description{
Forms both Gamma matrix and u vector
}
