% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbsfs-poly.R
\name{polyhedra.cbsfs}
\alias{polyhedra.cbsfs}
\title{Polyhedra for circular binary segmentation fixed steps}
\usage{
\method{polyhedra}{cbsfs}(obj, record.nrows = TRUE, ...)
}
\arguments{
\item{obj}{cbsfs object}

\item{record.nrows}{Whether or not to record the row number of the Gamma
matrix from each step of the algorithm.}

\item{...}{void, not used}
}
\value{
a polyhedra
}
\description{
The polyhedra has elements \code{gamma} (a matrix) and \code{u}
(a vector).
}
