% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbsft-poly.R
\name{polyhedra.cbsft}
\alias{polyhedra.cbsft}
\title{Polyhedra for circular binary segmentation fixed threshold}
\usage{
\method{polyhedra}{cbsft}(obj, ...)
}
\arguments{
\item{obj}{cbsft object}

\item{...}{void, not used}
}
\value{
a polyhedra
}
\description{
The polyhedra has elements \code{gamma} (a matrix) and \code{u}
(a vector).
}
