% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addpv.R
\name{polyhedra.fl}
\alias{polyhedra.fl}
\title{Helper to harvest polyhedra from FL object.}
\usage{
\method{polyhedra}{fl}(obj, numSteps = NULL, record.nrows = TRUE)
}
\description{
Helper to harvest polyhedra from FL object.
}
