% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flasso_polyhedra.R
\name{polyhedra.flFs}
\alias{polyhedra.flFs}
\title{Generate polyhedra matrix from flFs object}
\usage{
\method{polyhedra}{flFs}(obj, ...)
}
\arguments{
\item{obj}{flFs object}

\item{...}{not used}
}
\value{
An object of class polyhedra
}
\description{
Forms both Gamma matrix and u vector
}

