% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyhedra.R
\name{polyhedra.matrix}
\alias{polyhedra.matrix}
\title{Generate polyhedra}
\usage{
\method{polyhedra}{matrix}(obj, u, ...)
}
\arguments{
\item{obj}{numeric matrix to be the gamma matrix}

\item{u}{numeric vector}

\item{...}{not used}
}
\value{
object of class polyhedra
}
\description{
Generate polyhedra
}
