% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyhedra.R
\name{polyhedra.path}
\alias{polyhedra.path}
\title{Obtains polyhedron from |path| object. A |path| object is created from
generalized lasso dual path algorithm in the |genlassoinf| R package.}
\usage{
\method{polyhedra}{path}(obj, numSteps = obj$maxSteps)
}
\description{
Obtains polyhedron from |path| object. A |path| object is created from
generalized lasso dual path algorithm in the |genlassoinf| R package.
}
