% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyhedra.R
\name{polyhedra}
\alias{polyhedra}
\title{Generate polyhedra}
\usage{
polyhedra(gamma, u)
}
\arguments{
\item{gamma}{numeric matrix}

\item{u}{numeric vector}
}
\value{
object of class polyhedra
}
\description{
Generate polyhedra
}

