% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbsfs-poly.R
\name{polyhedra.wbsfs}
\alias{polyhedra.wbsfs}
\title{Polyhedra export function.}
\usage{
\method{polyhedra}{wbsfs}(obj, numSteps = NA, record.nrows = TRUE)
}
\arguments{
\item{obj}{Object of class |wbsfs|}

\item{numSteps}{Number of steps desired.}
}
\value{
object of class |polyhedra|.
}
\description{
Polyhedra export function.
}
