% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbsft-poly.R
\name{polyhedra.wbsft}
\alias{polyhedra.wbsft}
\title{Generate polyhedra matrix from wbs output
Forms both Gamma matrix and u vector}
\usage{
\method{polyhedra}{wbsft}(obj, ...)
}
\arguments{
\item{obj}{Output from wbs}

\item{...}{not use now}
}
\value{
An object of class polyhedra
}
\description{
Generate polyhedra matrix from wbs output
Forms both Gamma matrix and u vector
}
