% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize.R
\name{polyhedra_fusedlasso}
\alias{polyhedra_fusedlasso}
\title{Function to generate |polyhedra| object from fused lasso path output, from a
|path| class object generated from the genlassoinf package.}
\usage{
polyhedra_fusedlasso(obj, v = NULL, reduce = FALSE, sigma = NULL,
  verbose = FALSE, ...)
}
\arguments{
\item{obj}{Output from fused lasso}

\item{v}{a contrast vector, if you want to use smart addition of polyhedra.}

\item{reduce}{If TRUE, then does a Vup/Vlo comparison to see if you should
add (chunks) of rows instead of /all/ of them.}

\item{sigma}{noise level.}

\item{verbose}{load or not.}
}
\value{
An object of class polyhedra
}
\description{
Function to generate |polyhedra| object from fused lasso path output, from a
|path| class object generated from the genlassoinf package.
}
