% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyhedra.R
\name{print.fl}
\alias{print.fl}
\title{Proprietary print object for |path| class object. This is temporary, and
assumes that fused lasso (and not a different form of generalized lasso) is
run.}
\usage{
print.fl(obj)
}
\description{
Proprietary print object for |path| class object. This is temporary, and
assumes that fused lasso (and not a different form of generalized lasso) is
run.
}
