% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildBinSeg_fixedStep.R
\name{print.wbsFs}
\alias{print.wbsFs}
\title{Print function for convenience, of |wbs| class object.}
\usage{
\method{print}{wbsFs}(obj)
}
\description{
Print function for convenience, of |wbs| class object.
}
