% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{printprogress}
\alias{printprogress}
\title{A helper function to print the progress of a simulation.}
\usage{
printprogress(
  isim,
  nsim,
  type = "simulation",
  lapsetime = NULL,
  lapsetimeunit = "seconds",
  start.time = NULL,
  fill = FALSE
)
}
\description{
A helper function to print the progress of a simulation.
}
