% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{prune_of_1_length_segments}
\alias{prune_of_1_length_segments}
\title{Only compare things on nodes whose start and end are >1}
\usage{
prune_of_1_length_segments(Tcurr, Scurr, Ecurr)
}
\description{
Only compare things on nodes whose start and end are >1
}
