% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue.R
\name{pvalue}
\alias{pvalue}
\title{P-values for post-selection inference}
\usage{
pvalue(y, polyhedra, contrast, sigma = 1, value = 0)
}
\arguments{
\item{y}{numeric vector}

\item{polyhedra}{polyhedra object}

\item{contrast}{contrast numeric vector}

\item{sigma}{numeric to denote the sd of the residuals}

\item{value}{the null-hypothesis mean to test against}
}
\value{
a numeric p-value between 0 and 1
}
\description{
P-values for post-selection inference
}

