% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_splitting.R
\name{pvalue_ss}
\alias{pvalue_ss}
\title{P-values for sample splitting}
\usage{
pvalue_ss(y, contrast, sigma = 1)
}
\arguments{
\item{y}{numeric vector}

\item{contrast}{contrast numeric vector}

\item{sigma}{numeric to denote the sd of the residuals}
}
\value{
a numeric p-value between 0 and 1
}
\description{
This is limited to one-sided p-values currently, and the null-mean
is set to be 0.
}
