% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{qqunif}
\alias{qqunif}
\title{Function to plot qqlot of p-values. Use extra parameter}
\usage{
qqunif(pp, main = NULL, plot.it = TRUE, cols = NULL, type = c("p",
  "l"), legend.location = "bottomright", ...)
}
\arguments{
\item{pp}{numeric vector of p-values.}

\item{main}{label to plot as main title.}

\item{cols}{colors if |pp| is a list of numeric vectors.}
}
\description{
Function to plot qqlot of p-values. Use extra parameter
}
