% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{qqunif_add}
\alias{qqunif_add}
\title{Function to /add/ qq plot points of p-values}
\usage{
qqunif_add(pp, main = NULL, ...)
}
\arguments{
\item{pp}{numeric vector of p-values.}

\item{main}{label to plot as main title.}

\item{...}{other parameters for \code{qqplot()}.}
}
\description{
Function to /add/ qq plot points of p-values
}
