% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize.R
\name{randomize_addnoise}
\alias{randomize_addnoise}
\title{Conduct noise-added saturated model inference, for fused lasso or binary
segmentation (But also applies to sequential changepoint methods that
creates a valid polyhedron and has $cp and $cp.sign, like fused lasso from
the genlassoinf::dualpathSvd2()). Implements an importance sampling scheme.}
\usage{
randomize_addnoise(
  y,
  sigma,
  sigma.add,
  v,
  orig.fudged.poly = NULL,
  numSteps = NA,
  numIS = 100,
  bits = 1000,
  orig.fudged.obj = NULL,
  ic.poly = NULL,
  max.numIS = 2000,
  inference.type = c("rows", "pre-multiply"),
  verbose = FALSE,
  min.num.things = 10,
  mc.cores = 1,
  start.time = NULL,
  warn = FALSE,
  stable.thresh = 0.001
)
}
\arguments{
\item{y}{Data}

\item{sigma}{Gaussian noise standard deviation in data around mean.}

\item{sigma.add}{Amount of Gaussian noise to add. NOTE if equal to zero,
then reverts back to a non-marginalized noise addition.}

\item{v}{Test contrast vector. This is assumed to be fixed in the model
selection event.}

\item{orig.fudged.poly}{Original polyhedron object (of class
\code{polyhedra}) from adding a single noise.}

\item{numSteps}{Number of steps originally taken. Defaults to \code{NA}.}

\item{numIS}{Number of importance sampling replicates originally
desired. The importance sampling continues until there are
\code{min.num.things} number of valid importance sampling draws.}

\item{ic.poly}{Polyhedron of the (2-rise) stopping event.}

\item{max.numIS}{Maximum number of importance sampling replicates to take.}

\item{inference.type}{If equal to "rows", then the calculation of TG
statistics is done without modification. If equal to "pre-multiply",
then internally Gy = \eqn{\Gamma y} and Gv = \eqn{\Gamma v}. This serves
two purposes: first, if each polyhedra is too big in the first place,
this helps. Secondly, it speeds up calculation considerably
(poly_pval_from_inner_products() does this work, given Gv and Gy.)}

\item{verbose}{If \code{TRUE}, the importance sampling progress}

\item{mc.cores}{Number of cores to use for importance sampling. Defaults to
\code{1}.}

\item{start.time}{Start time of running time; usually output from
\code{Sys.time()}.}
}
\description{
Conduct noise-added saturated model inference, for fused lasso or binary
segmentation (But also applies to sequential changepoint methods that
creates a valid polyhedron and has $cp and $cp.sign, like fused lasso from
the genlassoinf::dualpathSvd2()). Implements an importance sampling scheme.
}
