% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize.R
\name{randomize_genlasso}
\alias{randomize_genlasso}
\title{Wrapper}
\usage{
randomize_genlasso(pathobj, sigma, sigma.add, v, orig.poly, numSteps = NULL,
  numIntervals, numIS, bits = NULL)
}
\arguments{
\item{pathobj}{An object from genlassoinf::dualPathSvd2(), of the class
"path".}

\item{sigma.add}{Standard deviation of noise, for additive noise.}

\item{v}{contrast vector}

\item{orig.poly}{original polyhedron to shift.}
}
\description{
Wrapper
}
