% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize.R
\name{randomize_wbsfs}
\alias{randomize_wbsfs}
\title{Conduct randomized inference for WBS. Implements an importance sampling
scheme.}
\usage{
randomize_wbsfs(
  v,
  winning.wbs.obj,
  numIS = 100,
  sigma,
  inference.type = c("rows", "pre-multiply"),
  cumsum.y = NULL,
  cumsum.v = NULL,
  stop.time = min(winning.wbs.obj$numSteps, length(winning.wbs.obj$cp)),
  ic.poly = NULL,
  bits = 50,
  max.numIS = 2000,
  min.num.things = 30,
  verbose = FALSE,
  mc.cores = 1,
  warn = FALSE,
  stable.thresh = 0.001,
  start.time = NULL
)
}
\arguments{
\item{v}{Test contrast vector. This is assumed to be fixed in the model
selection event.}

\item{winning.wbs.obj}{Winning WBS object of class \code{wbsfs}.}

\item{numIS}{Number of importance sampling replicates originally
desired. The importance sampling continues until there are
\code{min.num.things} number of valid importance sampling draws.}

\item{sigma}{Gaussian noise standard deviation in data around mean.}

\item{inference.type}{If equal to "rows", then the calculation of TG
statistics is done without modification. If equal to "pre-multiply",
then internally Gy = \eqn{\Gamma y} and Gv = \eqn{\Gamma v}. This serves
two purposes: first, if each polyhedra is too big in the first place,
this helps. Secondly, it speeds up calculation considerably
(poly_pval_from_inner_products() does this work, given Gv and Gy.)}

\item{ic.poly}{Polyhedron of the (2-rise) stopping event.}

\item{bits}{Number of precision bits to use for the calculation of the
Gaussian probabilities (regarding Vup and Vlo and vty).}

\item{max.numIS}{Maximum number of importance sampling replicates to take.}

\item{verbose}{If \code{TRUE}, the importance sampling progress is printed.}

\item{mc.cores}{Number of cores to use for importance sampling. Defaults to
\code{1}.}

\item{start.time}{Start time of running time; usually output from
\code{Sys.time()}.}

\item{y}{Data}

\item{orig.fudged.poly}{Original polyhedron object (of class
\code{polyhedra}) from adding a single noise.}

\item{numSteps}{Number of steps originally taken. Defaults to \code{NA}.}
}
\description{
Conduct randomized inference for WBS. Implements an importance sampling
scheme.
}
