% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize.R
\name{randomized_genlasso_pv}
\alias{randomized_genlasso_pv}
\title{Function for doing many noise-added fused lassos and computing randomized
p-value, from the beginning. It is more desirable to use the wrapper
randomized_genlasso(); this function will probably be retired not far from
now.}
\usage{
randomized_genlasso_pv(y, sigma, shift, sigma.add, D, v, orig.poly,
  numSteps = NULL, numIntervals, nsim.is, bits = NULL, reduce = FALSE,
  augment = TRUE)
}
\arguments{
\item{y}{data vector}

\item{sigma}{standard deviation of noise}

\item{D}{penalty matrix.}

\item{v}{Fixed contrast, formed /only/ with the knowledge of the selection
event on \code{y} with some fixed interval, and not from any more
information about \code{y}.}

\item{numSteps}{number of steps to take.}

\item{numIntervals}{number of WBS intervals you want /each time/. This
should match what you used when applying wild binary segmentation on
your observed dataset.}

\item{nsim.is}{Number of importance sampling samples you'd like to
calculate.}

\item{reduce}{\code{TRUE} if reduced version of polyhedron collecting is to
be used, in polyhedra collecting functions for WBS.}

\item{augment}{\code{TRUE} if WBS-FS should be run in augment mode.}

\item{v}{Contrast vector.}
}
\description{
Function for doing many noise-added fused lassos and computing randomized
p-value, from the beginning. It is more desirable to use the wrapper
randomized_genlasso(); this function will probably be retired not far from
now.
}
