% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildBinSeg_helper.R
\name{randomized_wildBinSeg_pv}
\alias{randomized_wildBinSeg_pv}
\title{Wrapper for doing many wild binary segmentations and computing p-values}
\usage{
randomized_wildBinSeg_pv(y, sigma, v, thresh = NULL, numSteps = NULL,
  numIntervals, nsim.is, bits = NULL)
}
\arguments{
\item{y}{data vector}

\item{sigma}{standard deviation of noise}

\item{v}{Fixed contrast, formed /only/ with the knowledge of the selection
event on \code{y} with some fixed interval, and not from any more
information about \code{y}.}

\item{thresh}{threshold.}

\item{numSteps}{number of steps to take.}

\item{numIntervals}{number of WBS intervals you want /each time/. This
should match what you used when applying wild binary segmentation on
your observed dataset.}

\item{nsim.is}{Number of importance sampling samples you'd
like to calculate.}
}
\description{
Wrapper for doing many wild binary segmentations and computing p-values
}
