% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-helper.R
\name{rejected}
\alias{rejected}
\title{Create verdicts from the vector of p-values (whose rownames are locations)
\code{pv} has location \code{loc}, accounting for how many tests were
conducted.}
\usage{
rejected(pv)
}
\arguments{
\item{pv}{numeric vector of p-values}

\item{loc}{location}
}
\description{
Create verdicts from the vector of p-values (whose rownames are locations)
\code{pv} has location \code{loc}, accounting for how many tests were
conducted.
}
