% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{restrict.intervals}
\alias{restrict.intervals}
\title{Sees which guys in object correspond to cusums calculated between |s| and
|e|. In other words, returns which guys are /qualified/.}
\usage{
\method{restrict}{intervals}(intervals, s = 1, e = n)
}
\description{
Sees which guys in object correspond to cusums calculated between |s| and
|e|. In other words, returns which guys are /qualified/.
}
