% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{rid.null}
\alias{rid.null}
\title{Gets rid of null elements in a list}
\usage{
\method{rid}{null}(mylist)
}
\arguments{
\item{mylist}{Some list that is suspected to contain some elements equal to NULL.}
}
\description{
Gets rid of null elements in a list
}

