% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{rid_jk_nicely_from_Tcurr}
\alias{rid_jk_nicely_from_Tcurr}
\title{Gets rid of the entries in the terminal node list Tcurr nicely,}
\usage{
rid_jk_nicely_from_Tcurr(Tcurr, Scurr, Ecurr, Tcurr.which.new)
}
\description{
Gets rid of the entries in the terminal node list Tcurr nicely,
}
\examples{
  Tcurr = list(c(1,1),c(1,2),NULL,NULL)
  Ecurr = Scurr = cplist(10)
  Ecurr = add(Ecurr, 1,1,7)
  Scurr = add(Scurr, 1,1,8)
  Ecurr = add(Ecurr, 1,2,11)
  Scurr = add(Scurr, 1,2,13)
  rid.jk.nicely(Tcurr,Ecurr,Scurr) ==  list(c(1,2),NULL,NULL,NULL)
}

