% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_helpers.R
\name{samplesplit}
\alias{samplesplit}
\title{Simple function to do all binary segmentation segment test after sample
splitting.}
\usage{
samplesplit(y, mn, numSteps, only.test.nulls)
}
\arguments{
\item{y}{data vector.}
}
\description{
Simple function to do all binary segmentation segment test after sample
splitting.
}
