% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{scale_resid}
\alias{scale_resid}
\title{Scale \code{resid} to have empirical std of \code{std}.}
\usage{
scale_resid(resid, std)
}
\arguments{
\item{resid}{numeric vector (meant to be residuals from changepoint model).}

\item{std}{desired empirical standard deviation}
}
\value{
Properly scaled \code{resid}.
}
\description{
Scale \code{resid} to have empirical std of \code{std}.
}
