% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyhedra.R
\name{smartadd}
\alias{smartadd}
\title{Adds new polyhedron to an original polyhedron, but only if the inference
changes (i.e. Vup and Vlo changes)}
\usage{
smartadd(orig.poly, new.poly, v, y)
}
\arguments{
\item{orig.poly}{Original polyhedron}

\item{new.poly}{New polyhedron to add, but only if the Vup or Vlo change.}

\item{v}{contrast vector}

\item{y}{data vector}
}
\description{
Adds new polyhedron to an original polyhedron, but only if the inference
changes (i.e. Vup and Vlo changes)
}
