% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyhedra.R
\name{snapshot.polyhedra}
\alias{snapshot.polyhedra}
\title{Gets a snapshot of the polyhedron from a particular algorithm step.}
\usage{
\method{snapshot}{polyhedra}(obj, numSteps)
}
\arguments{
\item{obj}{\code{polyhedra} class object.}

\item{numSteps}{Algorithm step you want a snapshot from}
}
\description{
Gets a snapshot of the polyhedron from a particular algorithm step.
}
