% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyhedra.R
\name{snapshot}
\alias{snapshot}
\title{Function generic, for a function to get a snapshot of the polyhedron from a
particular algorithm step.}
\usage{
snapshot(obj, ...)
}
\arguments{
\item{obj}{\code{polyhedra} class object.}
}
\description{
Function generic, for a function to get a snapshot of the polyhedron from a
particular algorithm step.
}
