% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{sqrt.mn.diff}
\alias{sqrt.mn.diff}
\title{Helper function to get right-to-left mean difference, or the contrast vector}
\usage{
\method{sqrt}{mn.diff}(s, b, e, n, y = NA, contrast.vec = FALSE,
  right.to.left = TRUE)
}
\description{
Helper function to get right-to-left mean difference, or the contrast vector
}

