% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize.R
\name{stable}
\alias{stable}
\title{Helper to discern whether a sequence of p-values have stabilized, based on
last two p-values in \code{p.so.far}.}
\usage{
stable(p, p.so.far, stable.thresh = 0.001)
}
\arguments{
\item{p}{new p-value under scrutiny (not included in p.so.far)}

\item{p.so.far}{numeric vector of p-values so far.}
}
\value{
TRUE or FALSE for whether p-values have stabilized or not.
}
\description{
Helper to discern whether a sequence of p-values have stabilized, based on
last two p-values in \code{p.so.far}.
}
