% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsfs.R
\name{summary.bsfs}
\alias{summary.bsfs}
\title{Summary of bsfs object}
\usage{
\method{summary}{bsfs}(object, ...)
}
\arguments{
\item{object}{bsfs object}

\item{...}{not used}
}
\value{
matrix of summary statistics
}
\description{
Summary of bsfs object
}
