% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flasso_fixedSteps.R
\name{summary.flFs}
\alias{summary.flFs}
\title{Summary for flFs object}
\usage{
\method{summary}{flFs}(object, ...)
}
\arguments{
\item{object}{flFs object}

\item{...}{not used}
}
\value{
summary of flFs (data frame)
}
\description{
Summary for flFs object
}

