% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_tree.R
\name{summary.Node}
\alias{summary.Node}
\title{Summary of Node object}
\usage{
\method{summary}{Node}(object, ...)
}
\arguments{
\item{object}{Node object}

\item{...}{not used}
}
\value{
matrix of summary statistics
}
\description{
Summary of Node object
}
