% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{t.statistic}
\alias{t.statistic}
\title{Calculates a t-statistic for E(v2)-E(v1) given vectors v1 and v2.}
\usage{
t.statistic(v1, v2)
}
\description{
Calculates a t-statistic for E(v2)-E(v1) given vectors v1 and v2.
}

