% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue.R
\name{tnorm.surv}
\alias{tnorm.surv}
\title{Temporarily added from selectiveInference package.}
\usage{
tnorm.surv(z, mean, sd, a, b, bits = NULL, correct.ends = TRUE)
}
\description{
Temporarily added from selectiveInference package.
}
