% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplist.R
\name{trim.list}
\alias{trim.list}
\title{Trims a list by deleting the last consecutive elements that are NULL.}
\usage{
trim.list(mylist, rid.null = FALSE)
}
\arguments{
\item{mylist}{Some list.}
}
\description{
Trims a list by deleting the last consecutive elements that are NULL.
}
