% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{trim.mat}
\alias{trim.mat}
\title{Function to trim a matrix from the right and bottom, ridding of all-NA rows/columns.
Returns NULL if mat is all NA's.}
\usage{
trim.mat(mat, type = c("rowcol", "row"))
}
\description{
Function to trim a matrix from the right and bottom, ridding of all-NA rows/columns.
Returns NULL if mat is all NA's.
}

