% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{trim.vec}
\alias{trim.vec}
\title{Trims a list by deleting the last consecutive elements that are NULL.}
\usage{
trim.vec(myvec)
}
\arguments{
\item{mylist}{Some list.}
}
\description{
Trims a list by deleting the last consecutive elements that are NULL.
}

