% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbsfs.R
\name{wbsfs}
\alias{wbsfs}
\title{Fixed step WBS function. There is a ``mimic'' option that helps you create
new polyhedron that says that the same |numSteps| winners win, but among a
new set of intervals.}
\usage{
wbsfs(
  y,
  numSteps,
  numIntervals = NULL,
  intervals = NULL,
  mimic = FALSE,
  wbs.obj = NULL,
  comprehensive = FALSE,
  cumsum.y = NULL,
  cumsum.v = NULL,
  inference.type = c("rows", "pre-multiply", "none"),
  stop.time = numSteps,
  ic.poly = NULL,
  sigma.add = NULL,
  v = NULL,
  precuts = NULL
)
}
\arguments{
\item{intervals}{Optionally, add an |intervals| class object that contains
intervals.}

\item{inference.type}{Specify whether you want to actually accumulate the
halfspaces, or just precalculate G*v and G*y for direct calculation of
p-values.}

\item{stop.time}{Manually inputted stop time. You can obtain a data
dependent stop time using local IC rises, via \code{ic_wrapper()}.}

\item{ic.poly}{polyhedron for the ic stopping event}

\item{v}{contrast vector; used only for ic.poly}
}
\value{
List of useful algorithm results, including the |gamma| matrix and
    |u|.
}
\description{
Fixed step WBS function. There is a ``mimic'' option that helps you create
new polyhedron that says that the same |numSteps| winners win, but among a
new set of intervals.
}
