% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbsft.R
\name{wbsft}
\alias{wbsft}
\title{Wild binary segmentation, with fixed threshold}
\usage{
wbsft(y, thresh, numIntervals = NULL, return.env = FALSE,
  seed = NULL, verbose = FALSE, intervals = NULL, augment = FALSE)
}
\arguments{
\item{y}{numeric vector to contain data}

\item{numIntervals}{number of random intervals to sample}

\item{numSteps}{numeric of number of stepsemas}

\item{tol}{tolerance to handle divide by zero instances}
}
\value{
either an environment address which contains output
    (\code{env$intervals} and \code{env$tree}), or a list of wild binary
    segmentation output.
}
\description{
Wild binary segmentation, with fixed threshold
}
