% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplist.R
\name{where_jk.cplist}
\alias{where_jk.cplist}
\title{Search in cplist$mat for the couplet (j,k) in the first two columns e.g. if
j=13 and k = 39, then it searches for the row (13,39,XXX) in an n by 3
matrix.}
\usage{
\method{where_jk}{cplist}(cplist, j, k, warn = FALSE)
}
\description{
Search in cplist$mat for the couplet (j,k) in the first two columns e.g. if
j=13 and k = 39, then it searches for the row (13,39,XXX) in an n by 3
matrix.
}
