% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{which.cross.precuts}
\alias{which.cross.precuts}
\title{Function to find which of the intervals (x[ii]:y[ii]) intersect (cross) with
any elements of |precuts|.}
\usage{
which.cross.precuts(x, y, precuts = NULL)
}
\description{
Function to find which of the intervals (x[ii]:y[ii]) intersect (cross) with
any elements of |precuts|.
}
