% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildBinSeg_fixedStep.R
\name{wildBinSeg_fixedSteps}
\alias{wildBinSeg_fixedSteps}
\title{Fixed step WBS function. There is a ``mimic'' option that helps you create
new polyhedron that says that the same |numSteps| winners win, but among a
new set of intervals.}
\usage{
wildBinSeg_fixedSteps(y, numSteps, numIntervals = NULL, intervals = NULL,
  mimic = FALSE, wbs.obj = NULL, comprehensive = FALSE)
}
\arguments{
\item{intervals}{Optionally, add an |intervals| class object that contains
intervals.}
}
\value{
List of useful algorithm results, including the |gamma| matrix and |u|.
}
\description{
Fixed step WBS function. There is a ``mimic'' option that helps you create
new polyhedron that says that the same |numSteps| winners win, but among a
new set of intervals.
}
