% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue.R
\name{ztest}
\alias{ztest}
\title{One-sided z-test regarding $v^Tmu$, using $v^Ty$.}
\usage{
ztest(y, v, sigma = 1)
}
\arguments{
\item{y}{data vector}

\item{v}{contrast vector}

\item{sigma}{noise standard deviation}
}
\description{
One-sided z-test regarding $v^Tmu$, using $v^Ty$.
}
