% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addpv.R
\name{addpv.cbsfs}
\alias{addpv.cbsfs}
\title{Appends the inference results to an object of class |bsFs|.}
\usage{
\method{addpv}{cbsfs}(
  obj,
  locs = NULL,
  type = c("plain", "addnoise"),
  sigma,
  sigma.add = NULL,
  declutter = FALSE,
  mn = NULL,
  min.num.things = 30,
  numIntervals = NULL,
  vlist = NULL,
  max.numIS = 2000,
  inference.type = c("rows", "pre-multiply")
)
}
\arguments{
\item{obj}{object of type bsFs}

\item{locs}{only test locations in \code{locs}.}

\item{type}{One of \code{ c("plain", "addnoise")}. If equal to
\code{"addnoise"}, then \code{sigma.add} needs to be provided.}

\item{sigma}{Noise level (standard deviation) of data.}

\item{sigma.add}{Additive noise. Defaults to NULL, in which case no additive
noise randomization inference is done.}

\item{mn}{original mean vector.}

\item{max.numIS}{Maximum number of importance sampling replicates to perform.}

\item{inference.type}{One of \code{c("rows","pre-multiply")}. Defaults to '
\code{"rows"}. Use \code{"pre-multiply"} if the polyhedron is too big
for memory.}
}
\value{
\code{obj} but with new list values "pvs" and "vlist".
}
\description{
Appends the inference results to an object of class |bsFs|.
}
