% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binseginf-package.R
\docType{package}
\name{binseginf-package}
\alias{binseginf-package}
\alias{binseginf}
\title{Post-selection inference for segmentation methods}
\description{
Post-selection inference for segmentation methods (binary segmentation, wild binary segmentation)
}
