% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyhedra.R
\name{combine.polyhedra}
\alias{combine.polyhedra}
\title{Combines several polyhedra to a single polyhedron}
\usage{
\method{combine}{polyhedra}(...)
}
\arguments{
\item{...}{polyhedra objects to add}
}
\description{
Combines several polyhedra to a single polyhedron
}
