% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_interval.R
\name{confidence_interval}
\alias{confidence_interval}
\title{Confidence interval from post-selection inference}
\usage{
confidence_interval(y, polyhedra, contrast, sigma = 1, alpha = 0.05,
  gridsize = 250)
}
\arguments{
\item{y}{numeric vector}

\item{polyhedra}{polyhedra object}

\item{contrast}{contrast numeric vector}

\item{sigma}{numeric to denote the sd of the residuals}

\item{alpha}{numeric between 0 and 1 with default of 0.05}

\item{gridsize}{numeric to denote how fine of a grid to invert the hypothesis
test}
}
\value{
a vector of two numbers, the lower and upper end of the confidence interval
}
\description{
Confidence interval from post-selection inference
}

