% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs-poly.R
\name{.gammaRows_from_comparisons}
\alias{.gammaRows_from_comparisons}
\title{Helper to make a set of rows (or multiplications with v)}
\usage{
.gammaRows_from_comparisons(vec, mat, sign.win, n, y = NULL, v = NULL,
  w = NULL, inference.type = c("rows", "pre-multiply"),
  excessive = FALSE)
}
\arguments{
\item{excessive}{Experimental feature that /additionally/ (and excessively)
conditions on the signs of competitor CUSUMs.}
}
\description{
Helper to make a set of rows (or multiplications with v)
}
