% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{getcusums}
\alias{getcusums}
\title{Get all cusums, given start point \code{s} and end point \code{e}}
\usage{
getcusums(s, e, y)
}
\arguments{
\item{s}{Starting index, between \code{1} and \code{n}}

\item{e}{Ending index, between \code{1} and \code{n}}

\item{y}{\code{n}-lengthed data vector.}
}
\value{
list of information about the cusum calculations and maximizers in
    this interval.
}
\description{
Get all cusums, given start point \code{s} and end point \code{e}
}
\examples{
n = 30
sigma = 1
lev1 = 0
lev2 = 5
mn = c(rep(lev1,n/3), rep(lev2,n/3), rep(lev1,n/3))
set.seed(0)
y = mn + rnorm(n,0,sigma)
getcusums(1,n,y)
getcusums(1,2*n/3,y)
}

