% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{halfspaces}
\alias{halfspaces}
\title{Calculates the halfspace vectors for the maximizing breakpoint and all the
signs.}
\usage{
halfspaces(s, b, e, thresh, n, y, is.terminal.node = F, verbose = F)
}
\arguments{
\item{is.terminal.node}{T/F for whether the node is one where the threshhold
is not breached.}
}
\description{
Calculates the halfspace vectors for the maximizing breakpoint and all the
signs.
}
\examples{
y = c(rnorm(10,0,1), rnorm(10,4,1))
# Calculate linear inequality vectors
##myineqs = halfspaces(s = 0, b = 10, e = 20, n = 20, y = y, type = "ineq")
}

