% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{halfspaces}
\alias{halfspaces}
\title{Calculates the halfspace vectors for the maximizing breakpoint and all the
signs, for fixed-threshold SBS.}
\usage{
halfspaces(s, b, e, z, thresh, n, y, is.terminal.node = F, verbose = F)
}
\arguments{
\item{is.terminal.node}{T/F for whether the node is one where the threshold
is not breached.}
}
\value{
A list of two objects \code{V} and \code{u}, for the halfspaces in
    represented as V'y>u
}
\description{
Calculates the halfspace vectors for the maximizing breakpoint and all the
signs, for fixed-threshold SBS.
}
