% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binSeg_fixedThresh.R
\name{is_valid.bsFt}
\alias{is_valid.bsFt}
\title{Checks of object is of class "bsFt", produced by binSeg_fixedThresh().}
\usage{
\method{is_valid}{bsFt}(obj)
}
\arguments{
\item{obj}{bsFt object}
}
\description{
Checks of object is of class "bsFt", produced by binSeg_fixedThresh().
}
