% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildBinSeg_fixedThresh.R
\name{is_valid.wbsFt}
\alias{is_valid.wbsFt}
\title{is_valid for wbs-ft}
\usage{
\method{is_valid}{wbsFt}(obj)
}
\arguments{
\item{obj}{wbsft object}
}
\value{
TRUE if valid
}
\description{
is_valid for wbs-ft
}
