% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circularBinSeg_fixedThres.R
\name{jumps.cbsFt}
\alias{jumps.cbsFt}
\title{Get jumps from cbsFt objects}
\usage{
\method{jumps}{cbsFt}(obj, ...)
}
\arguments{
\item{obj}{cbs object}

\item{...}{not used}
}
\value{
vector of all the jump locations
}
\description{
Enumerates the jumps for circular binary segmentation.
}
