% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{make_all_segment_contrasts}
\alias{make_all_segment_contrasts}
\title{Helper function for making segment contrasts from a wildBinSeg object OR
bsFt object or path object (from genlassoinf package), or just cps}
\usage{
make_all_segment_contrasts(obj)
}
\arguments{
\item{obj}{Result from running wbs()}
}
\description{
Helper function for making segment contrasts from a wildBinSeg object OR
bsFt object or path object (from genlassoinf package), or just cps
}
