% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{make_all_segment_contrasts_from_cp}
\alias{make_all_segment_contrasts_from_cp}
\title{Take |cp| and |cp.sign| and make all segment contrasts.}
\usage{
make_all_segment_contrasts_from_cp(cp, cp.sign, n,
  scaletype = c("segmentmean", "unitnorm"))
}
\arguments{
\item{cp}{integer vector of changepoints.}

\item{cp.sign}{integer vector of changepoint signs.}

\item{n}{length of data.}
}
\description{
Take |cp| and |cp.sign| and make all segment contrasts.
}
