% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binSeg_polyhedra.R
\name{polyhedra.bsFt}
\alias{polyhedra.bsFt}
\title{Function to collect polyhedra given some output from the
binSeg_fixedThresh(), of class "bsFt".}
\usage{
\method{polyhedra}{bsFt}(obj, verbose = F)
}
\arguments{
\item{obj}{object of bsFt class}

\item{verbose}{Whether or not to print things.}
}
\value{
An object of class polyhedra
}
\description{
Function to collect polyhedra given some output from the
binSeg_fixedThresh(), of class "bsFt".
}
