% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circularBinSeg_polyhedra_fixedSteps.R
\name{polyhedra.cbsFs}
\alias{polyhedra.cbsFs}
\title{Polyhedra for circular binary segmentation fixed steps}
\usage{
\method{polyhedra}{cbsFs}(obj, ...)
}
\arguments{
\item{obj}{cbsFs object}

\item{...}{void, not used}
}
\value{
a polyhedra
}
\description{
The polyhedra has elements \code{gamma} (a matrix) and \code{u}
(a vector).
}
