% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circularBinSeg_polyhedra_fixedThres.R
\name{polyhedra.cbsFt}
\alias{polyhedra.cbsFt}
\title{Polyhedra for circular binary segmentation fixed threshold}
\usage{
\method{polyhedra}{cbsFt}(obj, ...)
}
\arguments{
\item{obj}{cbsFt object}

\item{...}{void, not used}
}
\value{
a polyhedra
}
\description{
The polyhedra has elements \code{gamma} (a matrix) and \code{u}
(a vector).
}
