% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue.R
\name{pvalue}
\alias{pvalue}
\title{P-values for post-selection inference}
\usage{
pvalue(y, polyhedra, contrast, sigma = 1, null_mean = 0,
  alternative = c("one.sided", "two.sided"), precBits = NA)
}
\arguments{
\item{y}{numeric vector}

\item{polyhedra}{polyhedra object}

\item{contrast}{contrast numeric vector}

\item{sigma}{numeric to denote the sd of the residuals}

\item{null_mean}{the null-hypothesis mean to test against}

\item{alternative}{string of either "one.sided" or "two.sided" for the
alternative. If one.sided, the alternative means the test statistic is positive.}

\item{precBits}{precision of Rmpfr}
}
\value{
a numeric p-value between 0 and 1
}
\description{
P-values for post-selection inference
}
