% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_splitting.R
\name{pvalue_ss}
\alias{pvalue_ss}
\title{P-values for post-selection inference}
\usage{
pvalue_ss(y, contrast)
}
\arguments{
\item{y}{numeric vector}

\item{contrast}{contrast numeric vector}
}
\value{
a numeric p-value between 0 and 1
}
\description{
P-values for post-selection inference
}
