% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize.R
\name{randomize_wbsfs}
\alias{randomize_wbsfs}
\title{Synopsis: randomization wrapper for WBS.
New PV information based on newly drawn |numIntervals| - |numSteps|
intervals}
\usage{
randomize_wbsfs(v, winning.wbs.obj, numIS = 100, sigma,
  comprehensive = FALSE)
}
\description{
Synopsis: randomization wrapper for WBS.
New PV information based on newly drawn |numIntervals| - |numSteps|
intervals
}
