% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize.R
\name{rerun_wbs}
\alias{rerun_wbs}
\title{Helper for WBSFT randomization, in essence. Rerun WBS to get /new/, single
set of denom and numers from a new TG statistic, which is calculated from
the /single/ new set of halfspaces that characterize the maximization of the
original |winning.wbs.obj| but among /different/ set of
|numIntervals|-|numSteps| spaces.}
\usage{
rerun_wbs(winning.wbs.obj, v, numIntervals, numSteps, sigma)
}
\arguments{
\item{winning.wbs.obj}{Original contrast. We call it winning because we will
extract only the winning locations and /those/ winners' enclosing
intervals.}

\item{v}{test contrast}
}
\value{
A data frame (single row), with "pv" and "weight".
}
\description{
Helper for WBSFT randomization, in essence. Rerun WBS to get /new/, single
set of denom and numers from a new TG statistic, which is calculated from
the /single/ new set of halfspaces that characterize the maximization of the
original |winning.wbs.obj| but among /different/ set of
|numIntervals|-|numSteps| spaces.
}
