% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_splitting.R
\name{sample_splitting}
\alias{sample_splitting}
\title{Sample splitting}
\usage{
sample_splitting(y, method, ...)
}
\arguments{
\item{y}{data vector}

\item{method}{estimator on the first split}

\item{...}{additional arguments}
}
\value{
ss object
}
\description{
Even points are for finding the changepoints. Odd points are used for testing.
}
