% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binSeg_fixedSteps.R
\name{summary.bsFs}
\alias{summary.bsFs}
\title{Summary of bsFs object}
\usage{
\method{summary}{bsFs}(object, ...)
}
\arguments{
\item{object}{bsFs object}

\item{...}{not used}
}
\value{
matrix of summary statistics
}
\description{
Summary of bsFs object
}

