% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplist.R
\name{trim.vec}
\alias{trim.vec}
\title{Trims a vector by deleting the last consecutive elements that are NULL.}
\usage{
trim.vec(myvec)
}
\arguments{
\item{myvec}{Some list.}
}
\description{
Trims a vector by deleting the last consecutive elements that are NULL.
}
