% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-helper.R
\name{ztest}
\alias{ztest}
\title{Naive inference}
\usage{
ztest(v, y, sigma, alpha = 0.05)
}
\arguments{
\item{v}{contrast}

\item{y}{data}

\item{sigma}{noise level}
}
\value{
p-value for one-sided test of \eqn{H_0: v^Ty=0}.
}
\description{
Naive inference
}
